136954 by sweettea on 2018/11/08 (pair: raeburn) [VDOUP-4] Remove unnecessary bio function.
bioIsZeroData() is currently highmem-unsafe -- to be safe to
use on a general bio, it would need to account for the possibility
that the backing storage is in highmem, and use bvec_k{,un}map_irq
to determine the backing pages. However, rolling it into dataKVIO
and calling it only on the constant-length dataBlock is a small
transformation that avoids these complications.

.../bio.h
.../bio.c
Removed bioIsZeroData().

.../bio.c
.../dataKVIO.c
Moved isAllZeros() to dataKVIO, adapting it to take a KVIO,
assert that VDO_BLOCK_SIZE is a multiple of 8 bytes, and removing
the fallback loop for less than 8 bytes remaining.

.../dataKVIO.c
Reordered to copy data in before setting the zero block
status in order to always be checking the zero block
status of the data block.

Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>