137051 by sclafani on 2018/11/13 (pair: sweettea) Merged change 137002 from //eng/vdo to //eng/linux-vdo 137002 by sclafani on 2018/11/09 (pair: sweettea) Fixed bug in statistics for partial I/Os [VDO-4248] [BZ 1594406]
Partial I/O statistics were being reported as non-zero when not in
512B mode, which should have been impossible. The problem was that
the size of the original bio was being queried after it had been
used, and using it changed the size to less than a full block.

There was also a potential incompleteness in the statistics from
using both size and alignment to decide a bio is partial when
creating a DataKVIO, but only using the size to decide to count it
as partial I/O for statistics. The two conditions are likely the
same in practice (except possibly on discards), but using the same
criteria for statistics just in case they differ simplifies the
fix for the first issue.

.../dataKVIO.c
Moved the statistics increments for incomining partial I/Os
to where we set the isPartial flag of the DataKVIO, using
that flag instead of the bio size as the criteria.
Used the isPartial flag instead of the bio size to decide whether
to increment the bios acknowledged stats for partial I/Os.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>