137125 by tomj on 2018/11/15 (pair: corwin) VDOUP-6 Begin removing the dedupeIndex virtual interface.
The dedupeIndex virtual interface has been useful during our
development, but is not necessary and will not be welcome in the
upstream kernel. The code is split between a dedupeIndex module and
a udsIndex module. This change begins merging the two modules into a
single module.

Because udsIndex.h is small and dedupeIndex.h is large, we merge them
together into dedupeIndex.h. And because dedupeIndex.c is small and
udsIndex.c is large, we merge them together into udsIndex.c. This
creates a mismatch in naming but will also make any potential merges
easier.

A future change will deal with the naming mismatch between
dedupeIndex.h and udsIndex.c.

.../dedupeIndex.h
Do not declare albireoTimeoutJiffies. It is now a static variable in
udsIndex.c, which contains all the references to it.
Do not declare getAlbireoTimeout. It is now a static method in
udsIndex.c, which contains the definition and all the references to
it.

.../udsIndex.c
Move all the code from dedupeIndex.c into udsIndex.c.
Merge the makeUDSIndex method into the makeDedupeIndex method.

.../dedupeIndex.c
Since all the code is now in udsIndex.c, Delete dedupeIndex.c.

.../udsIndex.h
Since the only method prototype declared here is now merged into its
only caller, delete udsIndex.h.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: corwin <corwin@redhat.com>