137158 by sweettea on 2018/11/16 (pair: msakai, corwin) Made launchFlush() issue flushes regardless of write policy.
There are two ways in the base code to issue VIOs to flush the
storage: launchFlush() and launchWriteMetadataVIOWithFlush(). The
latter accepts flags about flush and FUA requests and issues flushes
in all cases because the underlying storage may require these flushes
for data integrity, whether in sync or async mode. launchFlush()
though only issues the flush in async mode, which is surprising and
inconsistent.

It turns out that all uses of launchFlush() are already conditional
on write policy, so removing the condition does not change the I/O
pattern of VDO. Removing the condition merely makes it less
likely that we later inadvertently try to use launchFlush()
for sync mode flushing, while adding an assertion preserves the
current intent to use it only in async mode (even if it would work
fine in sync mode).

Also, this removes a use of isFlushRequired(), which mildly furthers
VDOUP-3.

.../vio.h
.../vio.c
Removed launchFlush() condition causing a flush to only actually
be issued in async mode, adding an assertion in case we add a
use in sync mode unintentionally.

Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>