137434 by tomj on 2018/12/11 (pair: sclafani) ALB-2991 Merge changes 137335 and 137404 to Homer. 137335 by tomj on 2018/12/04 (pair: sclafani) ALB-2991 udsFlushBlockContext waits for the chapter writer to be idle
VDO is changing to call udsFlushBlockContext as part of suspending the
VDO device.  While VDO is suspended we should not be doing any I/O to
the index storage device.  Calling udsFlushBlockContext has always
assured that no UDS requests are still in progress.  And if no requests
are in progress we will not initiate any new index I/O (even for
checkpoints).  Except for the chapter writer thread, which writes out
the completed chapters asynchronously from the rest of UDS.

In this change we add to udsFlushBlockContext the guarantee that all
the asynchronous chapter writes have already completed.  The chapter
writer contains enough information to tell whether it will do any more
writing, and contains a mutex to protect the information, and even
broadcasts to a condition variable when it has finished doing its
writing.

This change makes a call to the udsFlushBlockContext method sufficient
for a "suspend without saving the index".  There is still work to be
done for "suspend with saving the index".

.../chapterWriter.h
.../chapterWriter.c
Implement the waitForIdleChapterWriter method, which needs to check a
boolean condition which already exists and only needs to be wrapped
inside an existing mutex plus condition variable.

.../indexRouter.h
.../localIndexRouter.c
.../remoteIndexRouter.c
Plumb waiting for the chapter writer to be idle though the IndexRouter
layer.  Local index routers need to pass the request on to the
chapter writer.  Remote index routers have no work to do and will
just return.

.../grid.h
.../grid.c
Plumb waiting for the chapter writer to be idle though the grid layer
to the IndexRouter layer.

.../context.c
Add code to the udsFlushBlockContext code path to wait for the grid to
have an idle chapter writer.

137404 by tomj on 2018/12/10 (pair: raeburn)
ALB-2991 Fix waitForIdleChapterWriter

Waiting for an idle chapter writer has been using the condition that we
are not idle when zonesToWrite == zoneCount.  This is correct in that
this condition means that we are either writing a chapter, or we have
signalled the chapter writer to wake up and write a chapter.  After the
chapter is written, zonesToWrite will be set to zero.

Unfortunately the condition is not entirely correct.  The correct
condition is zonesToWrite > 0.  When zonesToWrite is incremented from
0 to 1, the zone will send a control message to all the other zones
requesting that all zones immediately request the opening of a new
chapter.  Therefore even when there are no requests in the pipeline,
zonesToWrite > 0 means that we will write another chapter.

.../chapterWriter.c
Change the condition that identifies an idle state in the
waitForIdleChapterWriter method.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>