137606 by corwin on 2018/12/18 (pair: jwiele) Merge 137285 to linux-vdo. 137285 by corwin on 2018/12/01 (pair: sweettea) Simplified slab depot operations [VDO-4402].
This change is a vast simplification of the allocator operation
handling in slab depot. It is a prelude to the introduction of a
suspend operation.

.../slabDepotInternals.h
.../slabDepot.c
Modified the handling of actions (releasing tail block locks, saving
the slab depot, etc.) to use only a single pair of state objects
and one completion to manage all of the operations.
Renamed 'subTaskCompletion' to 'scrubbingCompletion' since it is now
only used to notify the VDO when scrubbing has completed.
Removed DepotResizeStep type and resizeStepRequested field.
Removed the save and action completions.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>