137651 by corwin on 2018/12/20 (pair: sweettea) Merged 137618 and 137621 to linux-vdo. 137618 by corwin on 2018/12/19 (pair: sweettea, sclafani) Separated read-only mode entry and allocation zone selection [VDO-4402]
This is the next step in cleaning up ThreadData before implementing
suspend/resume for VDO. It pulls the state and method for selecting
allocation zones into a new AllocationSelector since this functionality
and that of read-only mode entry are completely unrelated. Doing so
avoids some difficult naming issues when trying to divide the read-only
mode state into VDO level and per-thread level data.

.../allocationSelector.h
.../allocationSelectorInternals.h
.../allocationSelector.c
.../types.h
Created new AllocationSelector for selecting which physcial zone to
allocate from.
Modified getNextAllocationZone() to return the zone number instead of
the physical zone itself, thereby eliminating the need for
allocationSelector.c to include vdoInternal.h (and at effectively no
cost since this method has only one caller so the zone lookup is
still only in one place).

.../threadData.h
.../threadDataInternals.h
.../threadData.c
Removed getNextAllocationZone() and the fields it used.

.../allocatingVIO.h
.../allocatingVIO.c
Modified allocateDataBlock() to take an AllocationSelector and reflect
the change to getNextAllocationZone().

.../logicalZone.h
.../logicalZone.c
Added an AllocationSelector to LogicalZone and a getter for it.

.../packerInternals.h
.../packer.c
Added an AllocationSelector to Packer.

.../packer.c
.../blockMapTree.c
.../vioWrite.c
Reflected changes to allocateDataBlock().

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>