137745 by tomj on 2019/01/07 (pair: msakai) ALB-2991 Merge change 137737 to Homer 137737 by tomj on 2019/01/04 (pair: msakai) ALB-2991 Create the udsSaveIndex method.
.../uds.h
.../indexSession.c
.../udsModule.c
Add the udsSaveIndex method so that the index can be saved without
closing the index session.
Remove the mention of the udsAttachGridIndex method from the userLinux
version of the udsCloseIndexSession documentation.

.../grid.h
.../grid.c
.../albserver.c
.../gridRemote.c
.../indexSession.c
Remove the argument that specifies whether saving is optional from the
saveAndFreeGrid method.  Now that method always saves the index and
always frees the Grid.
Create the new saveGrid method that always saves the index and never
frees the Grid.

.../indexRouter.h
.../localIndexRouter.c
.../remoteIndexRouter.c
Add the argument that specifies whether freeing is optional to the
saveAndFree IndexRouter method.  Now that method optionally saves the
index and optionally frees the IndexRouter.

.../localIndexRouter.h
.../localIndexRouter.c
Add a needToSave flag to the LocalIndexRouter, which is set to true
when the next request to save the index should perform the save, or
is set to false when we already have the current state of the index
on the backing store.

137748 by bjohnston on 2019/01/08 (pair: joe)
[VDO-4405] Merge change 137387 to linux-vdo.
137387 by bjohnston on 2018/12/07 (pair: sweettea)
[VDO-4402] Add no flush flag for suspend.

.../kernelLayer.h
Add flag to inform suspendKernelLayer when to flush
out all our metadata and when not to.

.../dmvdo.c
In preresume, check dm flag and set our own kernel
flag. We need to do this because the dm flag is
reset before postsuspend is called.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>