137759 by bjohnston on 2019/01/10 (pair: sweettea) [VDO-4405] Merge changeset 137435 to linux-vdo. 137435 by bjohnston on 2018/12/11 (pair: sweettea) [VDO-4405] Go into read only when resume fails.
In grow cases, we have an active and inactive table
and a device config associated with each (the
inactive table was created during a dmsetup reload
and contains the new table line info). On a call
to dmsetup resume, the device mapper ioctl and table
code will call suspend on our device (if not already
suspended) then swap the tables and then call resume
on our device and finally call the destructor on the
old table. If our preresume code fails, the device
mapper code ignores any error and still calls the
destructor of the old table and so we have to make a
best effort at cleaning up.

.../dmvdo.c
If we get an error during modifyKernelLayer(), there is
not a whole we can do. We're in a pretty bad situation
given the device mappper code doesn't unwind the resumes
if one device returns an error. So the best we can do is
go into read only mode and swap the active config to the
new table info so we don't try to access freed memory
when the old table gets destroyed.

Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>