137800 by corwin on 2019/01/14 (pair: msakai) Merged 137720 to linux-vdo. 137720 by corwin on 2019/01/03 (pair: sweettea) Created ReadOnlyNotifier [VDO-4402].
This change is the next step in cleaning up the handling of read-only
mode as part of the suspend/resume work. It creates a new object, the
ReadOnlyNotifier, which replaces the ReadOnlyModeContext. It also
serves as a single object to hold what is left of ThreadData, all of
which is concerned with managing the entry to read-only mode.

.../threadData.h
.../threadDataInternals.h
.../threadData.c
.../readOnlyNotifier.h
.../readOnlyNotifierInternals.h
.../readOnlyNotifier.c
.../types.h
Moved threadData.* to readOnlyNotifier.*.
Modified constructor and destructor to create a single object, the
ReadOnlyNotifier which holds an array of ThreadData structures
instead of just having the ThreadData array.
Added a registration interface to ReadOnlyNotifier so that objects
which need to be notified when the VDO goes read-only can register
with the notifier. Added a 'listeners' field to ThreadData in order
to hold the list of listners which have registered for read-only
notifications on that thread.
Replaced the explicit knowledge in makeThreadReadOnly() that the
recovery journal and block allocators need to be notified with a
mechanism to notify any listeners which have registered with the
current thread.
Renamed setReadOnlyState() to startNotification() and removed the
saving of the super block and the setting of the VDO state. This
is now done via a listener which the VDO will register with the
notifier.
Renamed makeVDOReadOnly() to enterReadOnlyMode() in order to replace
the method from the ReadOnlyModeContext. Now that the
ReadOnlyNotifier is a distinct object (intead of just being a
ThreadData array), it can replace the context.
Reimplemented isReadOnly() to replace the check which was done on the
ReadOnlyModeContext.
Added a completion as the first field of ReadOnlyNotifier. This
completion is used by waitUntilNotEnteringReadOnlyMode() instead of
allocating one when the method is called. Modified
waitUntilNotEnteringReadOnlyMode() to only be callable from the admin
thread (which is the only thread it was ever called from in the base
code) so that launching the new completion is safe.

.../types.h
Removed ThreadData typedef.

.../readOnlyModeContext.h
.../readOnlyModeContextInternals.h
.../readOnlyModeContext.c
.../types.h
Removed ReadOnlyModeContext.

.../kernelVDO.c
Removed inclusion of threadData.h.

.../blockAllocator.h
.../blockAllocator.c
Modified notifyBlockAllocatorOfReadOnlyMode() to be a ReadOnlyListener,
made it static, and added registration of it with the
ReadOnlyNotifier when constructing a BlockAllocator.

.../recoveryJournal.h
.../recoveryJournal.c
Modified notifyRecoveryJournalOfReadOnlyMode() to be a
ReadOnlyListener, made it static, and added registration of it with
the ReadOnlyNotifier when constructing a RecoveryJournal.

.../vdo.h
.../vdoInternal.h
.../vdo.c
Exposed makeVDOReadOnly() and modified it to take a VDO instead of
a context since it is now only called from kernelVDO.c which has a
VDO. It is no longer used by ReadOnlyModeContext.
Added notifyVDOOfReadOnlyMode(), a read-only listener which sets the
VDO state to read-only and saves the super block.
Added enableReadOnlyEntry() which registers notifyVDOOfReadOnlyMode()
with the VDO's ReadOnlyNotifier.
Removed getThreadData() at it no longer has callers.
Removed isReadOnly() (which is now called directly on the notifier)
and enterReadOnlyModeFromContext() (since there are no more
contexts).

.../superBlock.c
Added an error handler for the super block write to log any save error,
in the case of a read-only mode save, this error was previously
logged by ThreadData, and other save errors were not consistently
logged.

.../upgrade.c
Modified upgradePriorVDO() to construct a ReadOnlyNotifier.

.../vdoLoad.c
Modified decodeVDO to construct a ReadOnlyNotifier for the VDO and then
call enableReadOnlyEntry().

.../blockAllocator.h
.../blockAllocatorInternals.h
.../blockAllocator.c
.../blockMap.h
.../blockMapInternals.h
.../blockMap.c
.../blockMapTree.h
.../blockMapTree.c
.../flush.c
.../packer.c
.../recoveryJournal.h
.../recoveryJournalInternals.h
.../recoveryJournal.c
.../recoveryJournalBlock.c
.../refCounts.h
.../refCountsInternals.h
.../refCounts.c
.../slab.c
.../slabDepot.h
.../slabDepotInternals.h
.../slabDepot.c
.../slabJournal.c
.../slabScrubber.h
.../slabScrubberInternals.h
.../slabScrubber.c
.../slabSummary.h
.../slabSummaryInternals.h
.../slabSummary.c
.../upgrade.c
.../vdo.h
.../vdoInternal.h
.../vdo.c
.../vdoClose.c
.../vdoLoad.c
.../vdoPageCache.h
.../vdoPageCacheInternals.h
.../vdoPageCache.c
.../vdoRecovery.c
.../vdoResize.c
.../vdoResizeLogical.c
.../vioWrite.c
.../vdoConfig.c
Reflected change from ReadOnlyModeContext to ReadOnlyNotifier.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>