137813 by bjohnston on 2019/01/15 (pair: sweettea) [VDO-4416] Merge change 137760 to linux-vdo. 137760 by bjohnston on 2019/01/10 (pair: sweettea) [VDO-4416] Only change version number when required.
.../dmvdo.c
Hard code the version number and only update it when
something changes that affects the user space tools
that work on VDO, like table line changes or things
like that. We are doing this so that LVM can use
dmsetup targets to get a verion and know how to
properly create and manage VDO properly.

Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>