138064 by tomj on 2019/02/19 (pair: msakai) VDO-4403 Merge changes 137265 and 137750 to //eng/linux-vdo 137265 by tomj on 2018/11/28 (pair: sclafani) VDO-4403 Flush the UDS request queue as part of VDO postsuspend.
This change addresses VDO-4403 by flushing the UDS request queues
at the end of the VDO postSuspend operation.  This change will also
contribute to the fix of ALB-2991 when UDS makes udsFlushBlockContext
also wait for any outstanding chapter writes to finish.

.../kernelLayer.c
At the end of suspending the kernel layer, flush the dedupe index.
This causes VDO to wait for all the outstanding dedupe requests
to complete.  We must wait to start this until after all the VIO
processing is complete.  There is no point is making this call
any earlier, as the flush operation is a passive operation that
simply waits for all the activity to be complete.

.../dedupeIndex.h
Create a flushDedupeIndex interface that invokes udsFlushBlockContext
on the dedupe index.

.../udsIndex.c
Implement the flush operation.
Fix processing of existing udsFlushBlockContext errors.  The only error
that can be returned from this method is an invalid block context id.

137750 by tomj on 2019/01/08 (pair: sweettea)
VDO-4402 Implement suspending/saving the UDS index.

This change implements what has been discussed and implied by the work
related to VDO-4402:  We should not need to do I/O to the backing store
after suspending the VDO device, because there is no guarantee that we
will see the device be resumed.  The device can shut down without
giving us the chance to write to it again.  So VDO suspend should
normally write out the UDS index as part of the suspend.

.../dedupeIndex.h
.../kernelLayer.c
Change the name of the flushDedupeIndex  method to suspendDedupeIndex,
which reflects what the method is really being asked to do.
Pass a flag to suspendDedupeIndex telling it whether we are doing a
suspend with save, or a suspend without save.

.../udsIndex.c
Implement the suspend operation that always calls udsFlushBlockContext
and optionally calls udsSaveIndex.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>