138123 by sclafani on 2019/02/27 (pair: msakai) Merged/adapted change 138077 from //eng/vdo to linux-vdo 138077 by sclafani on 2019/02/20 (pair: msakai) Worked around a compiler issue in our systemtap hooks [VDO-4441]
The asm() hack we use to try to keep parameters to the systemtap
hooks from being optimized away by GCC has stopped working for the
GCC version in RHEL8. A slight tweak would fix it, but it's in
shipped code, we need it on the Aluminum branch, and it's likely
not worthy of a BZ. This change adds a VDO_INTERNAL-only copy of
the asm() with the tweaked spelling, which we can merge to
Aluminum without affecting shipped code.

.../dataKVIO.c
Added a copy of the asm() hack in the systemtap hooks that uses
"m" instead of "g" to prevent optimization.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>