138289 by corwin on 2019/03/14 (pair: sclafani) Merged 137877 to linux-vdo. 137877 by corwin on 2019/01/23 (pair: msakai, sweettea) Converted object pool to new drain model for suspend/resume [VDO-4402].
When the suspend/resume work is complete, there will be four different
operations which involve some form of wait for some component or
components to drain their I/O. Much of the plumbing and scaffolding
for these operations factors nicely into a set of drain methods. This
change is the first step towards that model, factoring close (which
will become save) and suspend into a single drain method for object
pools. As part of this conversion, it is convenient to change
AdminState from an enum to a structure which holds both the enum and
a pointer to a completion waiting on the operation indicated by the
state.

.../adminState.h
.../adminState.c
Renamed the AdminState enum to AdminStateCode. Made AdminState a
struct which has both an AdminStateCode and a completion pointer.
Added SAVING, SAVED, and SUSPENDING state codes. Once all of the
suspend resume work is done, SAVING/SAVED will mean: "make the
on-disk state clean and then suspend", whereas
SUSPENDING/SUSPENDED will mean "drain I/O and persist enough metadata
to ensure no writes will be lost from a crash".
Removed CLOSING and CLOSED states and isClosing() since they are no
longer used.
Added adminState.c
Added predicates for the new state codes.
Added startDraining(), finishDraining(), finishDrainingWithResult(),
and resumeIfQuiescent() to provide common support for all drain
operations.

.../completion.h
.../completion.c
Added releaseCompletion() and releaseCompletionWithResult() as sugar
for copying a completion out of a structure, NULLing out the
reference to the completion in the structure, and then finishing the
completion.

.../objectPool.h
.../objectPool.c
Renamed suspendObjectPool() to drainObjectPool() which now is also
used for saving as well as suspending.
Removed openObjectPool(), closeObjectPool(), and the error return
in acquireEntryFromObjectPool() if closed since the concept of
closing is being replaced with suspending or saving.
Removed 'completion' field since it is now part of AdminState. Reworked
implementation both to support new admin state semantics and to use
the new admin state methods.

.../blockAllocator.c
.../blockMap.c
.../blockMapTree.c
Reflected changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>