138332 by corwin on 2019/03/15 (pair: msakai) Merged 137966 to linux-vdo. 137966 by corwin on 2019/02/06 (pair: sclafani) Made the page cache aware of the block map [VDO-4402], [VDOUP-6].
This change is a prequel to implementing suspend and resume in the
block map. By making the page cache aware of the block map zone in
which it lives, the drain methods become much cleaner and easier to
spell.

Furthermore, this change (mostly) eliminates an unused layer of
abstraction between the page cache and block map since the page cache
does not turn out to have other clients. Removing that abstraction is
one of the things we know is necessary for upstreaming. It remains
to explore what to do about the read and write hook function pointers
and the page contexts. Replacing those pointers with direct function
calls has severe implications for VDOPageCache_t1 and so should be
considered at a later date.

.../vdoPageCache.h
.../vdoPageCacheInternals.h
.../vdoPageCache.c
Replaced the threadID, readOnlyNotifier, and clientContext fields
with a pointer to the BlockMapZone which owns the cache.
Replaced the ReadHook and WriteHook signatures to explicitly take a
BlockMapZone instead of the (void *) clientContext.

.../blockMapInternals.h
Moved the 'readOnlyNotifier' field from BlockMapTreeZone to
BlockMapZone since it can now be shared both by the trees and the
page cache.

.../blockMapTree.h
.../blockMapTree.c
Removed 'readOnlyNotifier' parameter from initializeTreeZone() since
the field is now in the map zone.

.../blockMapTree.c
.../blockMap.c
Reflected changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>