138775 by tomj on 2019/04/12 (pair: sclafani) [VDOUP-8] Merge change 138506 to Jasper 138506 by sweettea on 2019/03/26 (pair: msakai) [VDOUP-8] Remove MurmurHash3_x64_128_double().
Nothing in the linux-uds tree uses this code, which is not present
in public code either, so we need not try to upstream it.

.../MurmurHash3.h
.../MurmurHash3.c
Removed MurmurHash3_x64_128_double().

138795 by tomj on 2019/04/15 (pair: sclafani)
UDSUP-6 Merge change 138518 to Jasper
138518 by tomj on 2019/03/27 (pair: sweettea)
Use the kernel doubly linked list facility

We currently have one doubly linked list in the UDS code, and it was
initially implemented with the user mode utilities from <sys/queue.h>.
Symmetrix supplied an equivalent version of <sys/queue.h> for us to
use, and in fact the Albireo and Symmetrix versions of <sys/queue.h>
were different versions of the same public source file from the 1990's.
When we ported to the Linux kernel, there was no such equivalent file,
so we settled on a single version (to simplify testing), and slightly
modified it so that it had no naming conflicts with the Linux kernel
code.

This change converts the UDS doubly linked list to use the interfaces
defined by the Linux kernel.  We provide an equivalent implementation
to be used by user mode UDS.

.../session.h
.../session.c
Convert the usage of the LIST macros defined by <sys/queue.h> to the
equivalent hlist facility defined by the Linux kernel.

.../kernelLinux/uds/queueDefs.h
Picks up the Linux kernel include file for use by the kernel build.

138811 by joe on 2019/04/16 (pair: msakai)
Reindent kernelVDOInternals.h (VDOUP-6).

.../kernelVDOInternals.h
Trivial change; reindent one line with tabs.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>