138812 by corwin on 2019/04/16 (pair: sclafani) Merged 138128 to linux-vdo. 138128 by corwin on 2019/02/27 (pair: msakai) Cleaned up AdminState [VDO-4402].
.../adminState.h
.../adminState.c
Converted AdminStateCodes to have structure instead of just being a
simple list of numbers. An AdminStateCode now consists of a type
(flush, save, etc), and some flags (draining, quiescent).
This structure makes the implementation of many predicates simpler
and makes adding new states possible without the need to update some
of the broader predicates (such as isDraining()).
Rewrote isDraining() and isQuiescent() to leverage the new structure.
Removed unused predicates.
Added isOperatingNormally().
Fixed a bug in startDraining() which was doing a NULL check on the
wrong completion when startDraining() is called when there is already
a waiter.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>