138818 by tomj on 2019/04/16 (pair: jwiele) VDO-4561 Change UDS index state before closing the index.
.../dedupeIndex.c
Before closing the UDS index, change the index state to IS_CHANGING,
so that getting the index statistics will not try to use the index
session.  This change removes the window where (1) the entries
indexed is zero and (2) the index is reported as offline but (3) the
new index is not created yet.  The index state will not return to
IS_OPENED until after the index is recreated or reopened.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: jwiele <jwiele@redhat.com>