138824 by tomj on 2019/04/16 (pair: msakai) UDSUP-1 Merge change 138552 to Jasper. 138552 by tomj on 2019/03/28 (pair: sweettea) Build the uds kernel module in using -std=gnu89
Actually, without specifying any -std option, so that we use the Linux
default which is -std=gnu89.

With this change, the UDS kernel module and the zubenelgenubi test
module build using the default compiler -std used by the Linux kernel.
The user mode only sources probably still contain some c99isms.

.../Makefile
Do not specify a -std option for compiling in kernel builds.

.../memoryLinuxKernel.c
.../threadsLinuxKernel.c
.../buffer.c
.../chapterIndex.c
.../deltaIndex.c
.../deltaMemory.c
.../errors.c
.../hashUtils.c
.../index.c
.../indexCheckpoint.c
.../indexComponent.c
.../indexInternals.c
.../indexLayout.c
.../indexLayoutParser.c
.../indexRouter.c
.../indexState.c
.../indexZone.c
.../logger.c
.../masterIndex005.c
.../masterIndex006.c
.../masterIndexOps.c
.../MurmurHash3.c
.../openChapter.c
.../openChapterZone.c
.../pageCache.c
.../parameter.c
.../recordPage.c
.../searchList.c
.../sparseCache.c
.../radixSort.c
.../volume.c
Fix all variable declarations embedded in a for statement.  This is the
only -std=c99 feature that we use UDS kernel code.

138830 by sclafani on 2019/04/17 (pair: corwin)
Merged vdoAudit change 138806 to linux-vdo [VDO-4449]
138806 by sclafani on 2019/04/15 (pair: sweettea)
Made vdoAudit warn when the VDO was not cleanly shut down [VDO-4449]

.../vdoState.h
.../vdoState.c
.../types.h
.../vdoInternal.h
.../vdo.c
Moved the declaration of VDOState to a new vdoState module.
Moved describeVDOState() to that module.

.../vdoState.h
.../vdoState.c
Created getVDOStateName() for use by vdoAudit.
Added VDO_STATE_COUNT to the state enum for getVDOStateName() to use.
Added explicit values to the VDOState enum constants since they
are stored in the volume and therefore must not change.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>