138854 by tomj on 2019/04/18 (pair: msakai) VDO-4546 Merge change 138774 to Jasper 138774 by tomj on 2019/04/12 (pair: jwiele) VDO-4546 Do not log every UDS request failure.
This is a reimplementation of change 138728 on //eng/linux-uds.

.../index.c
Remove the logUnrecoverable calls from the dispatchIndexZoneRequest
method.  These are the calls that send a flood of messages to
kern.log.  Instead, call the makeUnrecoverable method so that
the index session will be marked disabled.

.../logger.h
Update the documentation on the logUnrecoverable method to reflect what
the code really does.

.../errors.h
.../errors.c
Create the isSuccessful method.  A result code of UDS_SUCCESS or
UDS_QUEUED is a successful result.
Make the makeUnrecoverable method an always inline method.
Provide doxygen comments for the methods related to UDS_UNRECOVERABLE.

.../indexRouter.c
.../logger.c
Use the isSuccessful method.

138886 by msakai on 2019/04/23 (pair: tomj)
Merge 138857 and 138860 to linux-vdo. [VDO-4518]
138857 by msakai on 2019/04/18 (pair: tomj)
Remove vdoMonitor and its use in tests. [VDO-4518]

.../poolSysfsStats.c
Converted to kernel style.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>