138998 by sweettea on 2019/05/01 (pair: msakai) Merge 138881 to linux-vdo. 138881 by sweettea on 2019/04/22 (pair: sclafani) [VDO-4562] [VDO-4553] Rate-limit certain messages.
.../dataKVIO.c
Ratelimit ack latency violation messages.

.../vdoPageCache.c
Corrected setPersistentError() to only log if the error is not read-only
and we aren't already read-only in this zone, as the comment clearly
intended, thus obviating ratelimiting this message.
Ratelimited 'failed to write page' messages.

.../Makefile
Made vdoPageCache build with gnu89. dataKVIO is only ratelimited
internally, so need not be built with gnu89.

Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>