139173 by tomj on 2019/05/20 (pair: msakai) Merge change 138911 to Jasper. 138911 by tomj on 2019/04/24 (pair: msakai) UDSUP-7 Do not remove the safety seal.
The safety seal was a mechanism used by albmigrate in the user space
Albireo world.  We no longer have albmigrate.  Therefore we do not need
to use the safety seal.

This removes a piece of the UDS index that does not need to be
converted to used dm-bufio.

.../udsMain.c
.../indexLayout.h
.../indexLayout.c
Remove the code that removes the safety seal.  There is no longer any
code that writes the safety seal, and no code that reads the safety
seal.

139187 by tomj on 2019/05/21 (pair: msakai)
Merge change 138928 to Jasper
138928 by tomj on 2019/04/25 (pair: sweettea)
Identify and rename unused UDS status codes

.../uds-error.h
.../errors.c
Rename the unused externally visible UDS status codes to an obviously
unused name.

139202 by tomj on 2019/05/22 (pair: sweettea)
Merge change 138936 to Jasper
138936 by tomj on 2019/04/25 (pair: msakai)
Eliminate the UDS_UNSUPPORTED status code.

There are several places where we look for a UDS_UNSUPPORTED return
code, but there is no place that actually returns that value.

.../index.c
.../ioRegion.c
Do not look for a UDS_UNSUPPORTED return value.

.../uds-error.h
.../errors.c
Rename the UDS_UNSUPPORTED error code to indicate its unused status.

139203 by tomj on 2019/05/22 (pair: sweettea)
Merge change 138978 to Jasper
138978 by tomj on 2019/04/30 (pair: joe)
Refactor assertions code to be governed by __KERNEL_

.../kernelLinux/uds/assertDefs.h
Move the contents of the assertDefs.h headers directly into
assertions.h, governed by the __KERNEL__ define.

139224 by tomj on 2019/05/23 (pair: msakai)
Merge change 139174 to Jasper
139174 by tomj on 2019/05/20 (pair: msakai)
Refactor the allocSprintf method to be governed by __KERNEL__

.../stringUtils.c
.../memoryAlloc.h
.../memoryLinuxKernel.c
Refactor the doPlatformVasprintf methods directly into the allocSprintf
method, using an #ifdef __KERNEL__.

139256 by tomj on 2019/05/24 (pair: jwiele)
Merge change 139175 to Jasper
139175 by tomj on 2019/05/20 (pair: msakai)
Refactor away stringDefs.h by using __KERNEL__

.../stringUtils.h
.../kernelLinux/uds/stringDefs.h
Refactor the stringDefs.h files into stringUtils.h, using an #ifdef
__KERNEL__.

139414 by tomj on 2019/06/03 (pair: sweettea)
Merge changes 139177 and 139267 to Jasper
139177 by tomj on 2019/05/20 (pair: msakai)
Refactor random code to be governed by __KERNEL__

.../random.h
.../kernelLinux/uds/randomDefs.h
Refactor the randomDefs.h contents directly into random.h, governed by
#ifdefing __KERNEL__.  Then delete randomDefs.h.

.../random.h
Move static inline methods that are only used by test code from
random.h to testPrototypes.h.

.../random.c
.../Makefile
Move the user mode version of the fillRandomly method into random.c.
Then delete randomLinuxUser.c.

.../nonce.c
Add an include stringUtils.h, which is no longer provided by random.h.

.../udsModule.c
There is no need to export the fillRandomly method, because it is
static inline.  And VDO only uses it for user mode tests.

139440 by tomj on 2019/06/04 (pair: msakai)
UDSUP-7 Merge change 139337 to Jasper
139337 by tomj on 2019/05/30 (pair: msakai)
UDSUP-7 Do not write the volume header page.

We already removed the code that reads this page and checks it for
consistency with the UdsConfiguration (in change 138632).  There is
no need to write the page, and we will not need to justify writing
this page when UDS is upstreamed.

Indices written by Jasper (and the upstreamed code base) will be
created with version 2.  These indices will be readable on Homer
(which is used on RHEL8.1) but not on Gloria (which is used by
RHEL7.7 and RHEL8.0).

.../indexInternals.c
Do not format the volume.  All it did was to write the volume header
page.  We no longer read that page so there is no need to write it.

.../volume.h
.../volume.c
.../udsModule.c
Delete the formatVolume method.

.../volumeInternals.h
.../volumeInternals.c
.../udsModule.c
Delete the encodeVolumeFormat method.
Delete constants used only by the volume header.

.../indexLayout.c
Bump the super block version number to 2.  The actual superblock
structure has not changed, but the version number change means that
an old UDS reading a new index will report an unsupported version.
Without this change, the old UDS would have reported a corrupted
index.
We had already removed support for storing multiple subindices on a
block device or in a single file.  So we can now require that any
index with a version 2 superblock must have only 1 subindex.

139429 by tomj on 2019/06/03 (pair: sweettea)
Delete the unused bufferOps module.

.../bufferOps.c
.../bufferOps.h
Delete these files, which are not used.

139558 by sclafani on 2019/06/10 (pair: sweettea)
Merged unaligned field change 139492 to linux-vdo [BZ 1718058]
139492 by sclafani on 2019/06/06 (pair: msakai)
Stopped passing addresses of unaligned fields in packed structures.

GCC 9 in Fedora 30 now issues warnings for this which are
converted to errors, breaking the build. These unaligned
references are technically incorrect usage that GCC had always
allowed and we were able to get away with on X86 because it mostly
doesn't care about alignment. [BZ 1718058]

.../blockMap.c
.../fixedLayout.c
.../header.c
.../recoveryJournal.c
.../slabDepot.c
.../vdo.c
.../volumeGeometry.c
.../parseUtils.c
Stopped passing addresses of unaligned fields to buffer decoding
functions, replacing them with local variables that are safely
copied back into the fields through the structure reference.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>