139608 by tomj on 2019/06/12 (pair: msakai) Merge change 139431 to Jasper 139431 by tomj on 2019/06/03 (pair: sclafani) Rename unused internal error codes.
.../errors.h
.../errors.c
Rename unused internal error status codes to make it obvious which
codes are unused.  Do not change the existing error numbers. In
theory, we could just collapse the errors into contiguous numbers,
but this is complicated by VDO having a copy of the error message
table and we want do not want to change the existing error numbers.
The VDO copy of the table needs to be cleaned up.

139609 by tomj on 2019/06/12 (pair: msakai)
Merge change 139554 to Jasper
139554 by tomj on 2019/06/10 (pair: msakai)
Remove the unused filename field from the IndexComponentInfo.

.../indexComponent.h
.../indexPageMap.c
.../indexState.c
.../indexStateData.c
.../masterIndexOps.c
.../openChapter.c
Remove the unused filename field from the IndexComponentInfo.

139610 by tomj on 2019/06/12 (pair: msakai)
UDSUP-6 Merge change 139560 to Jasper
139560 by tomj on 2019/06/10 (pair: sweettea)
UDSUP-6 Replace the usage STAILQ, which is only available in user mode.

UDS makes one use of a STAILQ, which is implemented in an old user mode
include file named <sys/queue.h>.  The Linux kernel does not have an
implementation of a STAILQ, and the user mode implementation does not
live up to the Linux kernel way of doing things.

The Linux kernel does use a similar data structure for its struct
bio_list.  But is is special purpose code for a list of bios.

Convert our usage of STAILQ into straight c code that will be
acceptable in our upstream code base.

.../request.h
.../pageCache.h
.../pageCache.c
.../volume.c
Convert the STAILQ link pointer to be a normal Request pointer.
Define a RequestList (like a struct bio_list), and convert the STAILQ
head to be a RequestList.
When taking a snapshot of the RequestList, just convert it to a simple
Request pointer.  We do not try to add any more entries to the list
after taking the snapshot.

139625 by tomj on 2019/06/12 (pair: msakai)
UDSUP-6 Merge changes 139561 and 139619 to Jasper
139561 by tomj on 2019/06/10 (pair: sweettea)
UDSUP-6 Convert queue include files with cleaner use of __KERNEL__.

The older implentation in kernel mode is a queue.h file that is just an
include of <linux/list.h>.  The kernel reviewers are unlikely to accept
that.  So we use __KERNEL__ to select whether we include the linux file
or our own user mode file.

.../session.h
Use __KERNEL__ to either include <linux/list.h> or hlist.h.

.../hlist.h
Move the implementation of struct hlist into this file for user mode.

.../queue.h
.../kernelLinux/uds/queueDefs.h
.../userLinux/uds/queueDefs.h
Delete these files, which are no longer used.

139619 by tomj on 2019/06/12 (pair: msakai)
Clean up ioFactory includes by using __KERNEL__.

.../ioFactory.h
.../kernelLinux/uds/ioFactoryDefs.h
.../userLinux/uds/ioFactoryDefs.h
Refactor the contents of the ioFactoryDefs.h files into ioFactory.h
using an ifdef of __KERNEL__.
Delete ioFactoryDefs.h.

139655 by sclafani on 2019/06/13 (pair: msakai)
Merged GitHub README change 139650 to linux-vdo.
139650 by sclafani on 2019/06/13 (pair: msakai)
Linked to Red Hat VDO documentation in the GitHub README files.

.../README.md
Replaced entry saying "documentation will be added later" with
links to the RHEL7 and RHEL8 VDO documentation.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>