139846 by tomj on 2019/06/28 (pair: msakai) Merge changes 139741, 139754 and 139809 to Jasper 139741 by tomj on 2019/06/20 (pair: sclafani) Remove export of openVolumeRegion for tests.
.../udsModule.c
Remove export of openVolumeRegion, which is no longer used by any
tests.

139754 by tomj on 2019/06/21 (pair: msakai)
Refactor the common code used for creating an IndexLayout.

.../indexLayoutLinuxKernel.c
Refactor the common code in the makeIndexLayout methods into the
makeIndexLayoutFromFactory method.  The job of makeIndexLayout
becomes converting the index name string into an IOFactory that
encapsulate the index storage, and then invoking the common code
in the makeIndexLayoutFromFactory method.

.../indexLayout.h
.../indexLayout.c
Create the makeIndexLayoutFromFactory method.
139809 by tomj on 2019/06/25 (pair: raeburn)
Fix the linux-uds kernel build on RHEL8

.../threadsLinuxKernel.c
Add include of <linux/completion.h>.

139920 by tomj on 2019/07/09 (pair: sweettea)
UDSUP-7 Merge change 139845 to Jasper
139845 by tomj on 2019/06/28 (pair: jwiele)
UDSUP-7 Delete readonly index support from linux-uds.

We do not use any readonly index, and there are some separate code
paths for reading data from a readonly volume.  It is easier to just
delete the unused code than to convert the code paths to use dm-bufio.

.../indexInternals.h
.../indexInternals.c
.../index.c
.../indexZone.h
.../indexZone.c
Delete the readOnly parameter to the allocateIndex and makeIndexZone
methods.  We always pass in !READ_ONLY_INDEX.

.../defines
.../udsModule.c
.../readOnlyVolume.c
.../readOnlyVolume.h
Delete the readOnlyVolume module.

139924 by tomj on 2019/07/09 (pair: sweettea)
Remove support for a readonly volume from UDS

.../Makefile
Do not ship or build the readOnlyVolume module.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>