139934 by corwin on 2019/07/09 (pair: msakai) Merged 138147 to linux-vdo [VDO-4402]. 138147 by corwin on 2019/03/01 (pair: msakai) Implemented suspend and resume for SlabJournal (part 1) [VDO-4402].
This change introduces an AdminState to SlabJournal. The step of
combining the existing flush and close methods into a drain method
requires some additional work in the slab completion to be able to
pass the operation type. This will come in a future change.

.../slabJournal.h
.../slabJournalInternals.h
.../slabJournal.c
Added resumeSlabJournal().
Factored the core of flushSlabJournal() and closeSlabJournal() into a
currently static drainSlabJournal() method.
Replaced flushState and closeRequested fields with an AdminState.
Reworked flush and close handling to use the AdminState.
Implemented suspend other than providing an entry point for it
(this will come later).

.../adminState.h
.../adminState.c
Added isQuiescing() as an additional flag to note which states are
quiescing states.
Rewrote finishDrainingWithResult() in terms of the new flag.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>