139960 by corwin on 2019/07/10 (pair: sweettea) Merged 138526 to linux-vdo. 138526 by corwin on 2019/03/27 (pair: sclafani, msakai, sweettea) Improved handling of enqueue errors for reference blocks [VDO-4402].
This is a small clean up which is a precursor to implementing suspend
and resume on the RefCounts.

.../refCounts.h
.../refCounts.c
Factored both of the enqueueWaiter() calls for dirty reference blocks
into a single enqueueDirtyBlock() method. This method ensures that the
VDO will always goes into read-only mode if the enqueue fails.
Moved several methods to the top of the file to avoid the need for
forward declarations.
Reordered clauses in finishReferenceBlockWrite() to avoid a second
read-only mode check.

.../refCounts.c
.../blockAllocator.c
.../slabJournal.c
.../slabRebuild.c
Reflected changes to dirtyBlock().

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>