139967 by corwin on 2019/07/10 (pair: sweettea) Merged 138711 to linux-vdo. 138711 by corwin on 2019/04/09 (pair: msakai, sweettea) Began elimination of SlabCompletion [VDO-4402].
This is the first in the final set of changes for implementing suspend
and resume of the slab depot. The current implementation has evolved
over time and involves a number of ad hoc mechanisms for each of the
administrative functions on each of the components.

The aim of this set of changes is to unify the idioms and mechanism for
performing adminstrative actions on the depot, ensuring that multiple
adminstrative actions are not in progress simultaneously, and providing
a unified mechanism for tracking the administrative state of the depot
and its components.

The overall goal is to replace the maze of twisty passages currently
known as the SlabCompletion with an iteration mechanism in the block
allocator itself. The overall state will be managed via the slab
depot's action manager along with AdminState objects in each allocator
and in each slab.

In this new world, the administrative state of a slab journal and its
reference counts are managed via the slab instead of separately. This
reflects the evolution of the slab journal and reference counts which
no longer have as complicated a set of interactions between each other
as they once did.

This change completes the unification of the slab depot load process
beugn in change 138558, allowing the removal of the loadSlabJournals()
from SlabCompletion.

.../slabDepot.h
.../slabDepotInternals.h
.../slabDepot.c
Modified loadSlabDepot() to use the depot's ActionManager. The summary
load is now the prequel of the load action, and the loading of slab
journal tails or the erasing of the journals is done by each block
allocator for that allocator's slabs. For clean loads, the allocation
of the reference counts for each slab is now handled by each slab
instead of doing an additonal iteration over them.
Removed the now unused 'slabCompletion' and 'loadWaiter' fields, and
NEW_LOAD SlabDepotLoadType.

.../adminState.h
.../adminState.c
Added new states for the 4 different flavors of load: formatting (new),
normal, recovery, and rebuild. Added a new flag for loading states.
Added start and finish methods for loads.
Refactored start and finish methods for load, drain, and operation.

.../slab.h
.../slab.c
Added an AdminState to Slab to allow managing the quiescence of each
slab as a whole, instead of the journal and reference counts doing
so individually.
Added loadSlab() and notifySlabJournalIsLoaded() to support loading
slab journal tails from the block allocator directly instead of via
the SlabCompletion; notifyslabJournalIsLoaded() will also allocate
slab reference counts on a clean load.

.../blockAllocator.h
.../blockAllocatorInternals.h
.../blockAllocator.c
Added loadBlockAllocator().
Added an AdminState to the block allocator.
Added support for a BlockAllocator to apply a method to each of its
slabs without using a SlabCompletion.

.../slabDepot.c
.../blockAllocator.c
Moved the erasing of slab journals when loading for read-only rebuild
from being done by the slab depot to being done by each block
allocator. This allows all of the journal loads to have the same
structure at the top level. It would be a cleaner spelling still to
have each slab journal do its own erasing, but this would entail
significantly more memory since the slab iteration is done in
parallel for all the slabs belonging to a given allocator.

.../slabCompletion.h
.../slabCompletion.c
Removed loadSlabJournals().

.../slabJournal.h
.../slabJournal.c
Modified decodeSlabJournal() to only take the journal as a parameter
and for finishDecodingJournal() (which was renamed from
finishManualIO()), the ultimate conclusion of decodeSlabJournal() to
call slabJournalLoaded() on the Slab since the Slab is now
responsible for handling the load completion.
Moved the check for not reading the slab journal from
readSlabJournalTail() to decodeSlabJournal() since there's no point
in getting a VIO from the pool for a journal we're not going to end
up reading.

.../slabSummary.h
.../slabSummaryInternals.h
.../slabSummary.c
Modified loadSlabSummary() to take an already prepared parent and an
operation instead of a callback and error handler; replaced the
summary's 'completion' field with a 'loadParent' field which is a
pointer to hold the parent. The summary's own completion, and the
VDOActions it used to take are unnecessary since the supplied actions
will always be finishParentCallback(). The operation is used to
determine whether to read the existing summary on disk, or just write
out a blank summary.

.../completion.h
.../completion.c
Removed SLAB_SUMMARY_COMPLETION.

.../actionManager.h
.../actionManager.c
Modified the preamble to be asynchronous in order to support
loadSlabDepot(). Split InititatorAction type into ActionPreamble,
which is asynchronous, and ActionConclusion, which is synchronous.

.../slabDepot.c
Converted prepareForTailBlockCommit() to be an ActionPreamble.

.../blockMap.c
Converted prepareForEraAdvance() to be an ActionPreamble.

.../readOnlyRebuild.c
.../vdoLoad.c
.../vdoRecovery.c
Reflected change to loadSlabDepot().

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>