139970 by tomj on 2019/07/10 (pair: msakai) UDSUP-7 Merge changes 139892 and 139903 to Jasper 139892 by tomj on 2019/07/05 (pair: sweettea) UDSUP-7 Pass the buffer address into encodeRecordPage.
This change is in preparation for using dm-bufio in the kernel, where
we will follow the sequence of (1) using dm_bufio_new to associate a
buffer with the record page, (2) filling the record page, (3) using
dm_bufio_mark_buffer_dirty and (4) dm_release to send the record page
into the volume.  The point is that the buffer is given to us by
dm-bufio instead of being a scratch page that we keep reusing.

.../recordPage.h
.../recordPage.c
.../volume.c
Make the page buffer used for encoding a record page an argument to
the encodeRecordPage method.

139903 by tomj on 2019/07/08 (pair: sweettea)
UDSUP-7 Refactor writing of volume pages preparing to use dm-bufio.

The first implementation of Albireo wrote the volume as its own file.
The file I/O interface made it convenient to deal with byte offsets
into the volume file.

When using dm-bufio, it treats the storage as an array of fixed sized
blocks and want to know the block number that we want to read or write.
For the volume region of the index, the block is what we call a "page",
which can be either an index page or a record page.

In this change, we refactor the conversion of the page number (which we
call physicalPage) into the byte offset so that the conversion happens
in the writeScratchPage method.  When we replace writeScratchPage with
code to use dm-bufio, we will have the physicalPage value to use as the
block number.

.../volume.h
.../volume.c
.../volumeUtils.c
Modify the writeScratchPage method to take the physical page number
of the chapter instead of the byte offset to the chapter.
Do the increment of the physical page number in the caller of the
writeScratchPage method.  This replaces the update of the byte
offset.
Modify the writeIndexPages and writeRecordPages methods to take the
physical page number of the chapter instead of the byte offset to
the chapter.

.../geometry.h
.../geometry.c
Delete the recordPageOffset field of the geometry, which we no longer
use.

139974 by corwin on 2019/07/10 (pair: msakai)
Merged 138764 to linux-vdo.
138764 by corwin on 2019/04/12 (pair: msakai)
Eliminated use of SlabCompletion for scrubbing [VDO-4402].

.../slabScrubberInternals.h
.../slabScrubber.c
Moved the methods for reading and applying the slab journal to the
refCounts from slabRebuild.[hc] into slabScrubber, its only user.
Modified the scrubber to manage the rest of the scrubbing process
directly instead of using a SlabRebuild completion.

.../adminState.h
.../adminState.c
Added ADMIN_STATE_SCRUBBING and ADMIN_STATE_SAVE_FOR_SCRUBBING to
represent the two operations associated with scrubbing a slab. The
former commits the slab journal and loads the reference counts. The
latter saves the rebuilt reference counts.
Fixed a bug in the implementation of isNormal() where it always
returned true.

.../refCounts.h
.../refCounts.c
Added drainRefCounts() as the entry point for the two scrubbing
actions. In subsequent changes, the close and save methods will
be merged into this entry point as well.
Renamed checkForIOComplete() to checkForDrainComplete() and began
the conversion to use the AdminState in the slab instead of the
close requested flag in the RefCounts itself.
Made loadReferenceBlocks() static as it is now called via
drainRefCounts().
Removed spurious comment and confusing indirection in
loadReferenceBlock() which stated that the slab's VIO pool was being
accessed from the "load thread." This has not been true since all
loading of reference blocks has been done through the slab scrubber.

.../slab.h
.../slab.c
Added drainSlab() to provide a common bottleneck for drain operations
which cover both the slab journals and reference counts without
need of a SlabCompletion().
Added notifySlabJournalIsDrained() and notifyRefCountsAreDrained()
which are called when a slab journal or RefCounts has finished
draining.
Added resumeSlab() to replace resumeSlabJournal() and the not-yet-added
resumeRefCounts().
Added finishScrubbingSlab() to handle the state change and queuing of
a rebuilt slab. Also moved the initial scrubbing state change into
drainSlab() so that all of the internal changes to Slab are now done
in slab.c instead of in slabRebuild.c.
Added isSlabOpen() and isSlabDraining().

.../blockAllocator.h
.../blockAllocator.c
Removed saveRebuiltSlab(), the scrubber now calls drainSlab() with
ADMIN_STATE_SAVE_FOR_SCRUBBING on the Slab directly.

.../slabCompletion.h
.../slabCompletion.c
Removed saveSlab() which was called by saveRebuildSlab() now that the
scrubber calls drainSlab() directly.
Modified doRefCountIO() to reset the slab's admin state in order to
support the methods which have not yet been coverted to use
drainSlab() instead of SlabCompletion.

.../slabJournal.h
.../slabJournalInternals.h
.../slabJournal.c
Renamed drainSlabJournal() to oldDrainSlabJournal() in order to delay
converting flushSlabJournal() and closeSlabJournal() until the next
change.
Added drainSlabJournal() which is called by drainSlab() and which no
longer needs to take callbacks or parents as that is now managed
by the Slab.
Removed the 'adminState' field from SlabJournal and converted the
journal to use the AdminState in the journal's Slab.
Removed resumeSlabJournal() in favor of resumeSlab().

.../slabRebuild.h
.../slabRebuild.c
Removed slabRebuild.[hc].

.../completion.h
.../completion.c
Removed SLAB_REBUILD_COMPLETION.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>