139995 by corwin on 2019/07/10 (pair: msakai) Merged 138791 to linux-vdo. 138791 by corwin on 2019/04/15 (pair: sclafani) Combined ObjectPool and VIOPool [VDO-4402].
This change has been a long time coming in the suspend resume
work. Early on in the process, we determined that since VIOPool was now
the only ObjectPool, there was no reason for them to exist
independently. Furthermore, we determined that draining VIOPools was
redundant with the rest of the suspend nee close process. As of change
138779, the last caller of drainObjectPool() is now gone, and so the
merging of VIOPool along with ObjectPool along with the removal of the
drain support can be done in one step.

.../objectPool.h
.../objectPool.c
.../vioPool.h
.../vioPool.c
.../types.h
Merged ObjectPool into VIOPool and eliminted ObjectPool as an
independent structure.
Removed drain support.

.../blockAllocatorInternals.h
.../blockAllocator.c
.../blockMap.c
.../blockMapInternals.h
.../blockMapTree.c
.../forest.c
Reflected changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>