139993 by tomj on 2019/07/10 (pair: msakai) UDSUP-8 Merge change 139908 to Jasper 139908 by tomj on 2019/07/08 (pair: msakai) UDSUP-8 Fix the naming of IndexSession and UdsIndexSession.
Change 139883 removed the Session module, but left us with the
confusing typedefs of IndexSession as "struct indexSession" and
UdsIndexSession as "struct indexSession *".  This change is the
needed cleanup.

We also take the leap to not using the typedefs for the index session
structure, which will be needed upstream and is encouraged by the
UDSUP-8 story.  Since this will be an externally visible name in the
Linux kernel, we rename the structure to uds_index_session.

.../opaqueTypes.h
Remove the typedef of IndexSession.

.../indexSession.h
Rename struct index_session to struct uds_index_session.  An opaque
pointer to this struct is part of the public UDS interface.

.../indexSession.h
.../indexSession.c
.../request.h
.../udsMain.c
Eliminate the usage of IndexSession.  Replace IndexSession with the
usage of "struct uds_index_session".

.../uds.h
.../uds-block.h
.../request.h
.../udsMain.c
Eliminate the usage of UdsIndexSession.  Replace UdsIndexSession with
the usage of "struct uds_index_session *".
Note that we leave the typedef of UdsIndexSession in uds.h with an XXX
comment.  This will make the merge to Jenkins painless.  Then we can
update the usage of UdsIndexSession in the VDO tree before deleting
the typedef.

140000 by corwin on 2019/07/10 (pair: sclafani)
Merged 138820 to linux-vdo.
138820 by corwin on 2019/04/16 (pair: sweettea)
Fixed bug when freeing a VDO with unscrubbed slabs [VDO-4574].

Change 138764 converted block allocator close to be a drain which uses
the new applyToAllSlabs() method. The implementation of
applyToAllSlabs() missed the fact that the SlabCompletion's equivalent
loop would unsplice each slab from whatever ring it might be on before
acting upon it. By not doing so, slabs which are still queued for
scrubbing when the scrubber is stopped have their ring freed out from
under them when the depot is destroyed.

.../blockAllocator.c
Added unsplicing of each slab's ring node before applying a method to
it in applyToSlabs().

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>