140078 by sclafani on 2019/07/16 (pair: msakai) Merged table reload fix in change 139825 to linux-vdo [BZ 1669960] 139825 by sclafani on 2019/06/26 (pair: sweettea) Fixed VDO-4433 vdoStatus() can cause segfaults [BZ 1669960]
Replacing an inactive table by issuing two consecutive "dmsetup
load" commands was destroying the DeviceConfig referenced by the
KernelLayer. A "dmsetup status" issued concurrently or
subsequently would follow the pointer and (depending on timing).

This change fixes the problem by replacing the count of the number
of DeviceConfigs sharing a KernelLayer with a ring of those
configs. If the DeviceConfig the layer references is destroyed and
there are other DeviceConfigs referencing the layer, the layer is
re-pointed at one of those other configs.

.../kernelLayer.h
.../kernelLayer.c
Replaced the configReferences reference count field in KernelLayer
with a ring of the DeviceConfigs that are sharing the layer.
Removed functions that incremented and decremented the reference count.

.../deviceConfig.h
.../deviceConfig.c
Added a RingNode to DeviceConfig for putting them on the layer's ring.
Made freeDeviceConfig() zero out the config just before freeing it,
which makes it much easier to detect the use-after-free bug.
Created a setDeviceConfigLayer() that adds the config to the layer's
ring when the config references the layer, and removes it when
it stops referencing the layer.

.../dmvdo.c
Replaced calls to the reference counting function with calls to
setDeviceConfigLayer().
In vdoDtr(), checked if the ring is empty instead of checking if
the reference count is zero.
In vdoDtr(), when the layer isn't being destroyed, made it check if
the config being destroyed is still referenced by the layer, and
if it is, replace the reference with a config still in the ring.
This fixes the bug.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>