140083 by corwin on 2019/07/17 (pair: msakai) Merged 139137 to linux-vdo. 139137 by corwin on 2019/05/14 (pair: sweettea) Implemented suspend and resume for logical zones [VDO-4402].
.../logicalZone.h
.../logicalZone.c
.../types.h
Created a LogicalZones type which represents a collection of logical
zones rather than merely having an array of LogicalZone
structures. The new object contains an ActionManager and the VDO
along with the array of zones.
Replaced makeLogicalZone() and freeLogicalZone() with
makeLogicalZones() and freeLogicalZones().
Added getLogicalZone() for use by the LogicalZones' action manager.
Replaced closeLogicalZone() to drainLogicalZones() and converted it to
use the LogicalZones' action manager.
Added resumeLogicalZones().

.../vdoInternal.h
Changed type of 'logicalZones' field from LogicalZone ** to
LogicalZones *.

.../dataVIO.c
.../flush.c
.../vdo.c
.../vdoClose.c
.../vdoLoad.c
Reflected changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>