140095 by corwin on 2019/07/17 (pair: sclafani) Merged 139144 to linux-vdo. 139144 by corwin on 2019/05/15 (pair: msakai) Implemented suspend and resume of the packer [VDO-4402].
.../packer.h
.../packerInternals.h
.../packer.c
Renamed closePacker() to drainPacker() and modified it to set the
admin state to SUSPENDING and then call writeAllNonEmptyBins() which
was factored out of flushPacker().
Reimplemented flushPacker() to merely call writeAllNonEmptyBins() if
the packer is in normal operation. Since this call will put all
DataVIOs waiting in input bins onto the outgoing queue, and since
flushPacker() is no longer called by drainPacker() (nee
closePacker()), this is sufficient. No state change is needed in the
packer.
Replaced the 'closeRequest', 'flushing', and 'closed' fields with an
AdminState.
Renamed checkFlushProgress() to checkForDrainComplete() and rewrote it
in terms of the Packer's new AdminState.
Modified attemptPacking() to check the AdminState instead of the
'flushing' and 'closed' fields.
Added resumePacker().

.../vdoClose.c
Reflected changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>