140099 by corwin on 2019/07/18 (pair: sclafani) Merged 139327 and 139329 to linux-vdo. 139327 by corwin on 2019/05/29 (pair: msakai) Ensure that the AdminCompletion is not used concurrently [VDO-4402].
.../adminCompletion.c
Added an atomic boolean 'busy' field to AdminCompletion and used it in
performAdminOperation() to ensure that an AdminCompletion is only
used for one operation at a time.

139329 by corwin on 2019/05/29 (pair: raeburn)
Fixed unsafe field access in AdminCompletion [VDO-4402].

.../adminCompletion.c
Removed unsafe 'type' field access in the log message reporting that
an AdminCompletion is being used concurrently.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>