140105 by tomj on 2019/07/18 (pair: sclafani) Merge changes 140037 and 140055 to Jasper 140037 by tomj on 2019/07/12 (pair: msakai) Delete the udsState module.
The udsState module used to have a vital purpose:  it was where we
initialized the SessionGroups for index sessions and index contexts.
But we do not use SessionGroups any more.

The only remaining call that is needed is the open the logger in user
mode.  We move that open call elsewhere and just delete the udsState
module.

.../uds.h
.../udsModule.c
Eliminate the udsShutdown mathod, which has no callers.  The only thing
that the udsShutdown method still does is close the logger in user
mode.

.../logger.h
Make the openLogger and closeLogger methods user mode only.

.../loggerLinuxKernel.c
Eliminate the openLogger and closeLogger methods, which are not used
in kernel mode.

.../udsMain.c
Do not call udsInitialize.  The call is no longer needed.

.../indexSession.c
.../request.c
.../udsState.h
.../udsState.c
Delete the udsState module.

140055 by tomj on 2019/07/15 (pair: msakai)
Delete featureDefs.h

We created featureDefs.h to hold the definitions of flags that
controlled the existence or nonexistence of some features in
our source code.

.../volume.c
.../zone.c
Just use possible definition of __KERNEL__ to indicate the possible
existence of an environment that can be accessed by calling the
getenv method.

.../kernelLinux/uds/featureDefs.h
Delete these files which are no longer needed.

140108 by sclafani on 2019/07/18 (pair: jshimkus)
Merged README.md formatting fix in change 140085 to linux-vdo
140085 by sclafani on 2019/07/17 (pair: msakai)
Fixed formatting of documentation links in the GitHub README files.

When I added doc links in change 139650, I put a line break
between the link text and the link URLs for readability (since the
URLs are very long). This worked fine in the Markdown previewer I
was using, but it doesn't actually work on GitHub.

.../README.md
Removed line breaks from doc links.

140113 by tomj on 2019/07/18 (pair: sclafani)
Fix linux-vdo to work with the latest Jasper

Change 140105 to Jasper deleted some files that appear in the VDO
release.

.../Makefile
Do not build the udsState module, which has been deleted from UDS.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: jshimkus <jshimkus@redhat.com>