140115 by tomj on 2019/07/18 (pair: sweettea) Merge change 140070 to Jasper 140070 by tomj on 2019/07/16 (pair: sweettea) Rename freeIndexLayout to putIndexLayout.
.../indexLayoutLinuxKernel.c
.../udsModule.c
.../indexInternals.c
.../indexLayout.c
.../indexLayout.h
.../udsMain.c
Rename freeIndexLayout to putIndexLayout.  The name putIndexLayout
follows the kernel practice of pairing get-put routines to increment
and decrement a reference count, freeing an object when the count
drops to zero.

140120 by tomj on 2019/07/19 (pair: sweettea)
Merge change 140088 to Jasper
140088 by tomj on 2019/07/17 (pair: msakai)
Clean up the readOnly field of a Volume.

.../volume.h
.../volumeInternals.h
.../volumeInternals.c
.../volume.c
Delete the readOnly field of Volume.
Delete the readOnly argument to the allocateVolume method.  The value
passed in is always !READ_ONLY_VOLUME, so the argument is not needed.
Delete the READ_ONLY_VOLUME constant.

140121 by tomj on 2019/07/19 (pair: sweettea)
Merge change 140089 to Jasper
140089 by tomj on 2019/07/17 (pair: sclafani)
Delete the UdsIndexSession typedef.

.../uds.h
Delete the typedef of UdsIndexSession, which no longer has any users.

140140 by tomj on 2019/07/22 (pair: sweettea)
UDSUP-7 Merge change 140124 to Jasper
140124 by tomj on 2019/07/19 (pair: sweettea)
UDSUP-7 Add a syncVolume method.

Kernel UDS has been writing the volume using 32K synchronous writes.
The dm-bufio service will provide a cache that can contain many 32K
pages, and it can bundle them together into a smaller number of larger
write requests.  We will need a method that will flush all the dirty
buffers out to storage, and that will be the syncVolume method.

.../udsModule.c
.../volume.h
.../volume.c
Add the syncVolume method.

.../volume.c
.../volumeUtils.c
Add a syncVolume method call at the completion of writing a chapter.

.../index.c
Remove the syncRegionContents call from the saveIndex method.  This
call does not have any effect now that we are calling the syncVolume
method at the end of writing each chapter.
Refactor the saveIndex and saveIndexComponents method into a single
method.

140167 by sclafani on 2019/07/23 (pair: msakai)
Merged change 140132 to linux-vdo [BZ 1717435] [VDO-4648]
140132 by sclafani on 2019/07/22 (pair: raeburn)
Eliminated backtrace on physical size mismatch [BZ 1717435] [VDO-4648]

The expected physical size comes from the dmsetup table line as
of change 136038 last year, and logging a backtrace on a bogus
user input value for it has been deemed to be unacceptable.

.../vdo.c
Used logError() instead of ASSERT() for checking the physical size
in validateVDOConfig(). Revised the error message to reflect
that, and to be more user-oriented.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>