140299 by corwin on 2019/08/08 (pair: sclafani) Merged 139384 to linux-vdo [BZ 1659303]. 139384 by corwin on 2019/05/31 (pair: msakai) Converted close to suspend and/or save [VDO-4402].
This change replaces vdoClose with vdoSuspend which will suspend a VDO
and optionally save all dirty metadata as well.

.../vdoSuspend.h
.../vdoSuspend.c
Created new suspend admin operation which uses the single-callback
state machine idiom rather than a chain of callbacks. This
transaction will suspend the VDO, optionally writing out all dirty
metadata.

.../vdoClose.h
.../vdoClose.c
Removed vdoClose.[hc]

.../kernelVDO.c
Modified stopKVDO() to call performVDOSuspend() instead of
performVDOClose(), using the kernel layer's 'noFlushSuspend' field to
determine whether to write dirty metadata or not.

.../adminCompletion.h
.../adminCompletion.c
Exposed the AdminCompletion declaration by moving it from the .c file
to the .h.
Changed makeAdminCompletion() and freeAdminCompletion() to
initializeAdminCompletion() and uninitializeAdminCompletion() since
the VDO can now have an AdminCompletion instead of an
AdminCompletion *.
Exposed adminCompletionFromSubTask().
Added a 'phase' field to AdminCompletion to allow operations to use the
phased idiom.
Added an errorHandler parameter to performAdminOperation().

.../types.h
Removed the forward declaration of AdminCompletion since the structure
definition is now exposed in adminCompletion.h.

.../vdoInternal.h
Converted the adminCompletion field to be an AdminCompletion instead of
an AdminCompletion *.
Added an AdminState to track when the VDO has been suspended or saved.
Removed the 'closeRequested' field since it is no longer needed to
prevent incorrect exiting of recovery mode when the scrubber is
stopped before scrubbing has completed.

.../vdo.c
Removed the vdo->closeRequested check from leaveRecoveryMode() since it
is no longer the case that stopping the slab scrubber prematurely
will call into this method, and the field has now gone away.
Reflected changes to AdminCompletion.

.../vdoLoad.c
.../vdoResize.c
.../vdoResizeLogical.c
Reflected changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>