140389 by corwin on 2019/08/15 (pair: sweettea) Merged 139672 to linux-vdo [BZ 1659303]. 139672 by corwin on 2019/06/14 (pair: msakai) Fixed VDO-4663: fixed assertion failures from suspend/resume changes.
The final fix for VDO-4402, change 139590 introduced an assertion that
we don't submit metadata writes to the kernel layer while the VDO is
suspended. Unfortunately, that assertion incorrectly fires when doing
writes during pre-resume.

.../kernelLayer.h
.../kernelLayer.c
Added LAYER_STARTING and LAYER_RESUMING KernelLayerStates, which are
needed to differentiate the two different cases where the VDO gets
resumed.
Modified modifyKernalLayer() which is called during pre-resume to
only attempt modification if the layer is SUSPENDED, in which case
it also sets the state to RESUMING. If the layer is RUNNING, this
method is a no-op, otherwise it returns an error.
Modified startKernelLayer() to set the layer state to STARTING before
loading the VDO so that the super block update and possible offline
recovery can write. The state is set to RUNNING after startKVDO()
returns successfully.
Modified freeKernelLayer() to treat the new layer states the same
as the SUSPENDED state as all three of these states don't need
to perform a suspend on the base VDO.

.../kvio.c
Updated assertion in kvdoSubmitMetadataVIO() to not fire if the
state is RUNNING, RESUMING, or STARTING.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>