140496 by tomj on 2019/08/22 (pair: sweettea) UDSUP-7 Merge change 140328 to Jasper 140328 by tomj on 2019/08/09 (pair: sweettea) UDSUP-7 Write all kernel data using dm-bufio
The primary change is to be doing all of our storage writes using the
dm-bufio utility.

The secondary change is to update the BufferedReader_t1 test to use
an IOFactory to make its buffered reader and to not use an IORegion
in kernel mode to write its data.

For VDO-4344, it is worth noting that the preliminary measurements show
a 5X speed up in saving the index due to this change.

.../bufferedWriter.h
.../bufferedWriter.c
Change the BufferedWriter code to do I/O using dm-bufio in kernel mode,
while continuing to use an IORegion in user mode.  Maintain as much
commonality in both modes as is reasonable.
Always use a 4KB buffer.  UDS will have a hard time working on multiple
architectures if each architecture tries to use its own buffer size.
Split the makeBufferedWriter into separate kernel and user mode
implementations.  The kernel mode version needs a dm_bufio_client and
needs to prevent reading off the end of the region.  The user mode
version continues to need an IORegion and allocates its own buffer.
Create the prepareNextBuffer and flushPreviousBuffer helper methods for
use in kernel mode.  Include the dory code so that tests can disable
the writing of data.
Remove the performance optimization of writing large chunks of data
directly from the user buffer.  This optimization was put in when
we used to write hundreds of megabytes of data directly out of the
master index.  We no longer do any large writes in production code.

.../ioFactoryLinuxKernel.c
Change the openBufferedWriter method to use the new kernel mode
makeBufferedWriter interface.

.../bufferedReader.h
.../bufferedReader.c
.../udsModule.c
Remove the performance optimization of reading large chunks of data
directly into the user buffer.  This optimization was put in when
we used to read hundreds of megabytes of data directly into the
master index.  We no longer do any large reads in production code.
Remove the support for file-like behavior.  We ask for data in 4K
chunks because UDS always operates on chunks that are a multiple
of 4k in size.
Remove the getBufferedReaderPosition and etBufferedReaderPosition
methods, which are no longer used.
Do not export the makeBufferedReader method.  There are no longer any
tests that use it.

140503 by tomj on 2019/08/22 (pair: jwiele)
UDSUP-7 Merge change 140374 to Jasper
140374 by tomj on 2019/08/14 (pair: jwiele)
UDSUP-7 Read all kernel data using dm-bufio

The primary change is to be doing all of our storage reads using the
dm-bufio utility.  There is a lot of secondary cleanup.

.../bufferedReader.h
.../bufferedReader.c
.../udsModule.c
Change the BufferedWriter code to do I/O using dm-bufio in kernel mode,
while continuing to use an IORegion in user mode.  Maintain as much
commonality in both modes as is reasonable.
Always use a 4KB buffer.  UDS will have a hard time working on multiple
architectures if each architecture tries to use its own buffer size.
Split the makeBufferedReader into separate kernel and user mode
implementations.  The kernel mode version needs a dm_bufio_client and
needs to prevent reading off the end of the region.  The user mode
version continues to need an IORegion and allocates its own buffer.
Create a positionReader helper method to read the correct block and set
the read pointer to an offset within that block.
Delete the readBufferedData method, which was used by only one test.
The readFromBufferedReader method supports the reading from a block
stream, which is all that UDS needs to use.
Use the dm-bufio prefetch to read a few blocks ahead of the current
position. Reading ahead 4 blocks makes opening an index about 2X
faster.

.../ioFactory.h
.../ioFactoryLinuxKernel.c
.../udsModule.c
Remove the makeIORegion method.  All kernel I/O uses dm-bufio now.
Rewrite the openBufferedReader method to make an dm_bufio_client and
use the new makeBufferedReader interface.

.../ioRegion.h
Remove the getRegionBestBufferSize method, and all of the code behind
it. The buffered reader was the last user of this method.  UDS
really wants to use 4K buffers on all platforms for portability
of the index across architectures.

.../numeric.h
.../numeric.c
Remove the greatestCommonDivisor and leastCommonMultiple methods, which
are no longer used.

.../bufferedWriter.h
.../indexLayout.h
.../indexLayoutLinuxKernel.c
.../volume.h
Clean up #include lines.

.../Makefile
.../linuxIORegion.c
.../linuxIORegion.h
Delete the linuxIORegion module.  All kernel I/O uses dm-bufio now.

140508 by tomj on 2019/08/22 (pair: jwiele)
Merge change 140410 to Jasper.
140410 by tomj on 2019/08/16 (pair: sclafani)
Fix a format for a size_t to use %zu instead of %lu.

.../memoryLinuxKernel.c
Fix a format for a size_t to use %zu instead of %lu.

140509 by tomj on 2019/08/22 (pair: jwiele)
UDSUP-7 Merge change 140384 to Jasper.
140384 by tomj on 2019/08/15 (pair: sclafani)
UDSUP-7 Move ioRegion.h to userLinux

.../src/uds/ioRegion.h
Move ioRegion.h to the userLinux tree.  It is no longer used by any
kernel code.

140553 by joe on 2019/08/27 (pair: sclafani)
Detypedef actionManager.c (VDOUP-6).

.../actionManager.c
Remove structure typedefs.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>