140726 by tomj on 2019/09/12 (pair: sweettea) UDSUP-7 Merge change 140678 to Jasper 140678 by tomj on 2019/09/09 (pair: sweettea) UDSUP-7 Create struct volume_store
This change encapsulates the storage behind the volume into the struct
volume_store.  It is the first part of moving a lot of kernel vs. user
mode conditional code into a single module.

.../volumeStore.h
.../volumeStore.c
Create struct volume_store to be the actual storage behind the volume,
and the various methods for manipulating the volume store.
There will be more code added to the volumeStore module.  Eventually
all the kernel vs. user mode conditionals should be confined to the
volumeStore module.
Create the closeVolumeStore method to close the volume store.  The code
came directly from the freeVolume method.
Create the openVolumeStore method to open the volume store.  The code
came directly from the allocateVolume method.
Create the prefetchVolumePages to prefetch pages from the volume store.
All prefetching of these pages is due to be changed to call this
method.
Create the syncVolumeStore method to synch the volume store.  The code
came directly from the syncVolume method.

.../volume.h
.../volume.c
.../volumeInternals.c
Convert the bufioClient and region fields of the volume to use struct
volume_store.
Delete the syncVolume method, which is replaced by the syncVolumeStore
method in the volumeStore module.
Use the closeVolumeStore method when freeing the volume.
Use the openVolumeStore method when allocating the volume.

.../udsModule.c
Rename the syncVolume method to syncVolumeStore.

140727 by tomj on 2019/09/12 (pair: sweettea)
Merge change 140684 to Jasper
140684 by tomj on 2019/09/10 (pair: sweettea)
Clean up the accessMode module.

When Albireo stored its data in files, we used IOAccessMode to control
how files were opened and accessed.  Now that we store the index in a
stream of 4K blocks, we make little use of it.

.../indexState.h
Move the definition of enum IOAccessMode into this module.  Only
define IO_READ and IO_WRITE in the enum.  Enum IOAccessMode is only
used in the interface to the getStateIndexStateBuffer method, and
only those 2 values are valid.

.../indexLayout.h
Do not #include accessMode.h.

.../accessMode.h
Delete this file.

140732 by tomj on 2019/09/12 (pair: sclafani)
Fix side effects of Jasper merges 140726 and 140727.

Add the volumeStore module to the kernel build.  This module was added
by merge 140726.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>