140910 by tomj on 2019/09/25 (pair: sweettea) UDSUP-6 Merge change 140829 to Jasper 140829 by tomj on 2019/09/20 (pair: sweettea) UDSUP-6 Reimplement RequestQueue to use struct wait_queue.
The Linux kernel provides a wait_queue facility to implement the basic
operation of a thread waiting for a condition.  Albireo implemented
this operation using an EventCount, which avoids locking as much as
possible, only using a semaphore when it is necessary to wait.  When
it uses this semaphore, its performance depends upon the ability to
wait for a semaphore with a timeout that has microsecond granularity.

It is not difficult to implement such a semaphore in Linux, but past
attempts to add such a semaphore to the kernel have met with failure.

But the good news is that the wait_queue facility has the same
properties that an EventCount has.  It avoids locking as much as
possible, has the same prepare/wait/cancel semantics, and it has a
timeout that uses a high resolution timer.

.../requestQueueKernel.c
Create a new implementation of RequestQueue for kernel mode.
The major difference from the user mode code is that the entire wait
loop is implemented in the wait_event and wait_event_hrtimeout
macros.  So we need to prepare the wait conditions before we know
whether there will be a wait required.  This is not a problem, but
it does make the code read differently.
Make a minor change to the timing constants.  On our PMI machines, we
max out our performance at 400 to 450 nanoseconds per chunk.  This
appears to be the best we can ask for using only a single test thread
creating hashes to be indexed.  At that rate, we can only create 25
requests in 10 microseconds.  With minimum batch size of 32 requests,
we will not never need the minimum timeout of 5 microseconds, and
even the default starting point of 10 microseconds is too small.  So
I changed the minimum to 10 microseconds and the default starting
point to 20 microseconds.  This measured as a tiny performance
improvement in kernel testing.  A similar change was not seen in user
mode testing.

.../userLinux/build/Makefile
Remove requestQueue.c from the list of UDS common source files.
Add requestQueueKernel.c to the kernel build, and requestQueueUser.c
to the user mode build.

140911 by tomj on 2019/09/25 (pair: sweettea)
ALB-3021 Merge change 140834 to Jasper
140834 by tomj on 2019/09/20 (pair: sweettea)
ALB-3021 Merge change 140707 to linux-uds
140707 by tomj on 2019/09/11 (pair: sweettea)
ALB-3021 Improve chapter index verification during rebuild

.../index.c
When rebuilding the IndexPageMap, verify that the chapter index pages
contain a consistent number of index lists.  Due to ALB-3021, it is
conceivable but very rare for an inconsistent number to happen.  The
resulting errors would be confusing.

140912 by tomj on 2019/09/25 (pair: sweettea)
Merge change 140835 to Jasper
140835 by tomj on 2019/09/20 (pair: sweettea)
Remove bio accounting from UDS.

UDS has counters for the current and peak number of struct bios used.
But there is nothing today that changes these counters, so they always
report using the value 0.

These counters are not appropriate for UDS, as UDS does not use any
struct bios.  VDO once cared about these counters.  If it should care
in the future, it would be more proper for VDO to keep and manage its
own counters.

.../memoryDefs.h
.../memoryLinuxKernel.c
.../udsModule.c
Delete the recordBioAlloc and recordBioFree methods.  These methods are
never called.
Remove the bio counters from the getMemoryStats method interface.
These values will always be zero.

.../sysfs.c
Reflect changes to the getMemoryStats method.

140914 by tomj on 2019/09/25 (pair: sweettea)
ALB-3021 Merge change 140841 to Jasper
140841 by tomj on 2019/09/20 (pair: sweettea)
ALB-3021 Merge change 140724 to linux-uds
140724 by tomj on 2019/09/12 (pair: sweettea)
ALB-3021 Refactor initializing a DeltaIndexPage

.../deltaIndex.c
Refactor the initializeDeltaIndexPage method, splitting out the
verifyDeltaIndexPage method.  This will be a preferred refactoring
when we have to read a chapter index page that may be either little
or big endian.  In that case, the initializeDeltaIndexPage method may
need to call the verifyDeltaIndexPage method twice.
Do less logging in verifyDeltaIndexPage.  The log messages, while
detailed and informative during development, are not of great use
to end users.  If we try to verify an index page using the wrong
endianness, they will put useless noise into the kernel log.  And
it is likely that the upstream community will disapprove of the
logging.
Do not skip the check of the expected nonce when the expected nonce is
zero.  We only used a zero nonce during testing or for the old-style
multi-file Albireo user mod index.
Remove the redundant check that the last list ends on the page, because
it is covered by the check that the last list ends while leaving room
on the page for the post-field guard bits.

140917 by joe on 2019/09/25 (pair: sweettea)
Detypedef blockMapTree.h/blockMapTreeInternals.h (VDOUP-6).

.../blockMapTree.h
.../blockMapInternals.h
Remove struct typedef(s).
Decamelcase the struct identifier(s).

.../blockMapTree.c
.../blockMapTreeInternals.h
.../forest.c
.../forest.h
Adapt to the above changes.

140923 by tomj on 2019/09/25 (pair: sweettea)
Fix side effects of Jasper merges 140910 and 140912.

.../memoryUsage.c
Modify the arguements to the getMemoryStats method to match the
interface changes in merge 140912.

.../Makefile
Reflect the renaming and moving of files in merge 140910.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>