141010 by tomj on 2019/10/04 (pair: sweettea) ALB-3021 Merge change 140980 to Jasper 140980 by tomj on 2019/10/02 (pair: sweettea) ALB-3021 New index version that writes little endian chapter headers
Bump the index super block version up to 3.  Version 2 (and older)
indices will write native endian chapter index headers.  New indices
will be version 3 and write little endian chapter index headers.

.../indexVersion.h
.../indexVersion.c
.../defines
Create the indexVersion module.
Define the SUPER_VERSION_* constants.
Create struct index_version, and the initializeIndexVersion method that
populates it correctly from the given super block version number.
Document the super block version number history.

.../indexLayout.h
.../indexLayout.c
Add a struct index_version to the IndexLayout, and create the
getIndexVersion method to retrieve it from the IndexLayout.
Use the SUPER_VERSION_ constants.
Always initialize the struct index_version to reflect that actual
index that is loaded or created.
When loading an existing index, always check that the number of
subindices is 1.  This check is correct because we no longer have
any code to support more than one subindex.

.../index.c
.../chapterWriter.h
.../chapterWriter.c
When making the ChapterWriter, pass the index version parameters in,
so that the OpenChapterIndex will write the correct version of the
chapter index headers.

.../volume.c
.../chapterIndex.h
.../chapterIndex.c
Store constants that affect every chapter index header that is written
in the OpenChapterIndex structure.  These include the volume nonce
and the chapter index header endianness flag.
Remove the volume nonce parameter to the the packOpenChapterIndexPage
method.

.../deltaIndex.h
.../deltaIndex.c
Make the packDeltaIndexPage method obey the flag that determines the
correct endianness of the DeltaPageHeader.  A true value writes
native endian values, and a false value write little endian values.

141015 by tomj on 2019/10/05 (pair: sweettea)
Fix side effects of Jasper merge 141010

.../Makefile
Add the indexVersion module to the build.  This module was added by
merge 141010.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>