141025 by corwin on 2019/10/07 (pair: msakai) Merged 140918 and 14093 to linux-vdo. 140918 by corwin on 2019/09/25 (pair: msakai, sweettea) Fixed recovery journal resume issues [VDO-4768] [BZ 1752893].
When saving the recovery journal, the journal's active block pointer
was not cleared, but the block to which it was pointing would be moved
from the active list to the free list. After resuming, problems would
ensue.

While investigating, it was also discovered that when opening the
journal after a recovery, we would create a 1 block hole in the
journal. If we managed to do so 256 times in a row in the same spot
(i.e. crashing at exactly the same physical journal block each time
around the journal), the recovery byte would wrap which could cause the
next recovery to be incorrect. The chance of this happening is
vanishingly small, but it happens to be fixed by this change in any
event.

Finally, it also became apparent that the check to ensure that the
sequence number has not exceeded 48 bits was not applied to all of the
places where the journal tail was set or advanced. This change
addresses that as well.

All of these issues are fixed by modifying the recovery journal to only
have an active block when it has entries to make or when the most
recent block is incomplete.

.../recoveryJournal.c
Factored the sequence number check and the setting of the journal's
tail field into a new setJournalTail() method.
Replaced all modifications of the journal's 'tail' field with calls to
setJournalTail().
Modified checkForDrainComplete() to call notifyCommitWaiters() when
cleaning up in read-only mode rather than cleaning up the active list
itself.
Eliminated uses of the journal's active block when it does not have
entries to add or commit:
- Modified recycleJournalBlock() to clear the journal's active block
pointer if the block being recycled was the active block.
- Modified checkForDrainComplete() to recycle the active block if
there is one (there should be at most one active block by the time
all of the other drain checks pass).
- Modified makeRecoveryJournal() to not set an active block.
- Modified initializeRecoveryJournalPostRecovery() and
decodeRecoveryJournal() to not initialize the active block.
- Modified advanceTail() to attempt to set an active block. Only if
this succeeds will the tail be incremented and block map era
advanced. The method was also changed to return a boolean indicating
whether or not the tail was advanced.
- Pulled setActiveBlock() into advanceTail() which is now its only
caller.
- Modified prepareToAssignEntry() to attempt to advance the tail if
the active block is full (journal->activeBlock being NULL is now
treated as full). Modified this method to take the journal rather
than a recovery block as a parameter since there may be no active
block when this method is called.
Lifted the writeBlock() call from the end of the loop in
recycleJournalBlock() up to notifyCommitWaiters() since the new
caller of recycleJournalBlock(), checkForDrainComplete(), should
never have a block to write.
Removed the tail advances in assignEntry() as they are now all handled
by prepareToAssignEntry(). If the entry being added fills the block,
the block will be written, but the tail will not be advanced until
the next entry is to be added to the journal.
Moved the read-only mode entry due to lack of decrement space from
assignEntriesFromQueue() into prepareToAssignEntry(). This prevents
spurious entry into read-only mode when there is no in memory entry
space but there is room on disk for decrement entries. While this
currently can not happen, if the number of VIOs ever exceeds that
entry space in the in-memory tail blocks, it could.
Reflected changes to prepareToAssignEntry() in assignEntriesFromQueue().

.../recoveryJournalBlock.h
Modified isRecoveryBlockFull() to return true if the block is NULL.

.../recoveryJournalBlock.c
Modified shouldCommit() to return false if the recovery block is
NULL. This can happen during a save because slab journal locks are
released after the journal has been drained.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>