141212 by tomj on 2019/10/24 (pair: sweettea) Merge change 141161 to Jasper 141161 by tomj on 2019/10/22 (pair: sweettea) Convert thread definitions to use #ifdef __KERNEL__.
.../threads.h
.../kernelLinux/uds/threadDefs.h
.../threadCondVar.h
Move the contents of the threadDefs.h header files into the threads.h
header file, using #ifdef __KERNEL__.  Because threads.h is one of
the 2 source files that includes threadDefs.h, this has low impact.
Move the contents of threadCondVar.h into threads.h.  Because the
common practise is for the code to #include threads.h, and for
threads.h to include threadCondVar.h, this has low impact.  This
movement was needed because threadCondVar.h was the other place that
included threadDefs.h, and it was the easiest way to avoid a cycle of
dependencies in the header files.

141228 by tomj on 2019/10/25 (pair: sweettea)
Merge change 141165 to Jasper
141165 by tomj on 2019/10/22 (pair: sclafani)
Remove the timeInISOFormat method.

.../timeUtils.h
.../timeUtilsLinuxKernel.c
Delete the unused timeInISOFormat method, which is not used.

141229 by tomj on 2019/10/25 (pair: sweettea)
Merge change 141179 to Jasper
141179 by tomj on 2019/10/23 (pair: sweettea)
Refactor logger includes to use #ifdef __KERNEL__

.../kernelLinux/uds/loggerDefs.h
.../logger.h
Move the contents of the loggerDefs.h files into logger.h, using #ifdef
__KERNEL__ as appropriate.  Then delete the loggerDefs.h files.

141230 by tomj on 2019/10/25 (pair: sweettea)
Merge change 141180 to Jasper
141180 by tomj on 2019/10/23 (pair: sweettea)
Refactor platform specific error support code.

.../errors.c
.../kernelLinux/uds/errorDefs.h
.../errorsLinuxKernel.c
Move the systemStringError method and its supporting data into the
common errors module, using #ifdef __KERNEL__ as appropriate.  Then
delete the platform specific error file.s

.../Makefile
Delete the kernel specific errors module.

141231 by tomj on 2019/10/25 (pair: sweettea)
Merge change 141181 to Jasper
141181 by tomj on 2019/10/23 (pair: sweettea)
Refactor the timedWaitCond timeout.

The timedWaitCond method was created as a wrapper around the
pthread_cond_timedwait method, and implemented the wrapped interface,
which specified the timeout as an absolute time stored in a struct
timespec.  This is only convenient in user mode unix implementations
that use pthreads.

We abstracted the struct timespec for absolute times into the AbsTime
typedef to make the time specification easier to deal with.  But it
turns out that the kernel really wants a relative timeout value.

This change refactors the timeout, changing it from an absolute time
(when to stop waiting) into a relative time (how long to wait for).
Thus the kernel implmentation has the timeout that it needs, and we can
let the user mode implementation convert the timeout into the absolute
time that it needs.  This latter conversion is especially convenient to
do in the CondVar code, because this code knows that pthreads expects
the absolute time to be specified using the real time clock.

.../threads.h
Modify the timedWaitCond method to take a timeout parameter instead of
an absolute time deadline.

.../threadCondVarLinuxKernel.c
Modify the waitCond method to not call timedWaitCond
Modify the timedWaitCond method to just pass the timeout through to the
eventCountWait method.

141247 by tomj on 2019/10/25 (pair: sweettea)
Merge change 141182 to Jasper
141182 by tomj on 2019/10/23 (pair: sweettea)
Remove support for kernels older than RHEL8 or FEDORA29

RHEL8 is based on Linux 4.18, and FEDORA29 is based on Linux 5.0, so we
can remove all conditional code for versions before Linux 4.18.

.../resourceUsageLinuxKernel.c
Remove conditional for code older than Linux 4.11.

.../memoryLinuxKernel.c
Remove conditional for code older than Linux 4.13.

.../logger.h
Remove conditional for code older than Linux 4.15.

141248 by tomj on 2019/10/25 (pair: sweettea)
Merge change 141183 to Jasper
141183 by tomj on 2019/10/23 (pair: msakai)
Remove atomic definitions needed when we tested on RHEL7.

.../atomicDefs.h
Remove atomic_* definitions that we needed when our RHEL7 lab machines
were running RHEL7.4.
Remove ACCESS_ONCE definition.  It is obsolete, and we no longer use
it.

141249 by tomj on 2019/10/25 (pair: sweettea)
Merge change 141216 to Jasper
141216 by tomj on 2019/10/24 (pair: jwiele)
Clean up the modules and headers related to timeUtils.

.../timeUtils.h
.../timeUtils.c
Import the contents of the timeDefs headers, using #ifdef __KERNEL__ as
appropriate.
Import the contents of the timeUtils<PLATFORM> source, using #ifdef
__KERNEL__ as appropriate.
Make the deltaTime method static, as it is only used by the futureTime
method.
Make the deltaTime and futureTime methods user-mode only, as they are
only need by the pthread interfaces for processing timeouts.
Make the timeDifference method static inline in kernel mode.
Make the isValidTIme method user-mode only, as it is only needed to
deal with failures of the clock_gettime system call.
Make the relTimeToString and sleepFor methods test only.

.../udsModule.c
Remove export of futureTime, which is used only in user mode.
Remove export of timeDifference, which is static inline in kernel mode.

.../kernelLinux/uds/timeDefs.h
Remove timeDefs.h.

.../kernelLinux/build/Makefile
Remove the timeUtilsLinuxKernel module.

141268 by tomj on 2019/10/26 (pair: sclafani)
Fix side effects of Jasper merge 141249 and unpin

.../Makefile
Delete the timeDefs.h header files, and the timeUtilsLinuxKernel and
timeUtilsLinuxUser modules.  These were deleted by merge 141249 to
Jasper.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>