141569 by sclafani on 2019/12/04 (pair: corwin) Merged change 141488 from linux-vdo to Jasper [UDSUP-6] [VDO-4829] 141488 by sclafani on 2019/11/20 (pair: raeburn, msakai) Used interruptible flavors of wait_event to avoid hung task warnings.
Most drivers are supposed to use the interruptible forms unless
doing such low-level stuff as to not be able to handle it. Since
this is used on on our own worker thread and not a user thread, it
doesn't really matter, other than the wait state assigned to the
thread/task, which decides whether or not the hung-task checks are
performed.

This should hopefully fix VDO-4829 (or at least reduce the hung
task warnings to the rate we had before change 140829).

.../requestQueueKernel.c
Switched to using wait_event_interruptible() and
wait_event_interruptible_hrtimeout() in requestQueueWorker().
Re-arranged the work loop so interrupts won't rouse the thread
from dormancy or be counted as part of a batch.

141608 by joe on 2019/12/06 (pair: sclafani)
Detypedef InputBin type (VDOUP-6).

.../packerInternals.h
.../types.h
Remove struct typedef(s).
Decamelcase the struct identifier(s).

.../dataVIO.h
.../packer.c
Adapt to the above changes.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>