142397 by corwin on 2020/02/12 (pair: sclafani ) Merged 141350 to linux-vdo. 141350 by corwin on 2019/11/04 (pair: sclafani) Fixed premature block map drain notifications [VDO-4800] [BZ 1766358].
When draining the block map, each tree zone is told to drain. When each
tree zone has drained, it then tells the page cache to drain. When the
page cache is drained, it notifies the block map that the zone is
drained. The actual AdminState which informs both the tree zone and
page cache that it is being drained lives in the block map zone and is
shared by both components. Unfortunately, if the last outstanding I/O
from the page cache happens to complete before the tree zone has
finished draining, the page cache will send the drain
notification. While it is difficult to produce such a condition, it is
possible if rare in unit tests resulting in occasional failures when
the tree zone issues writes to an already quiescent layer. Although we
have seen no evidence of this issue causing problems in the kernel,
such issues can not be ruled out.

This change fixes the issue by modifying the drain complete check to be
done in a manner which checks that both components are drained before
sending the notification.

.../blockMapInternals.h
.../blockMap.c
Added checkForDrainComplete() which finishes the drain in a zone when
both the tree zone and the page cache drained.
Added an initiateDrain() method which kicks off a drain of both the
tree zone and the page cache and then checks whether the drain is
complete. Used this new method as the intiator for a zone drain
rather than draining the tree zone first. Since we now check both
components explicitly, we don't need to wait to start the cache drain
until the tree zone is done (especially since that didn't work as
spelled anyway).

.../blockMapTree.h
.../blockMapTree.c
Renamed checkForDrainComplete() to isZoneTreeActive(), exposed it, and
modified it to merely return a boolean indicating whether the zone is
active (i.e. not done draining) rather than draining the page
cache.
Replaced all calls to the former tree zone checkForDrainComplete() with
calls to the new block map method of the same name.
Converted drainZoneTrees() to take the tree zone now that it is no
longer being used as an AdminInitiator.
Removed the call to checkForDrainComplete() from drainZoneTrees() since
this call could trigger a premature drain finish in the event of a
save when the tree zone is inactive, and the page cache has no
outstanding I/O but does have dirty pages.

.../vdoPageCache.h
.../vdoPageCache.c
Renamed checkForIOComplete() to isPageCacheActive(), exposed it, and
modified it to merely return a boolean indicating whether the cache
is active (i.e. not done draining) rather than directly finishing the
drain.
Replaced all calls to checkForIOComplete() with calls to the new block
map method checkForDrainComplete().
Removed the call to checkForDrainComplete() from drainVDOPageCache().
While this call currently does no harm, it is redundant with the call
in blockMap.c:initiateDrain(). Furthermore, were the order of the
calls to drainZoneTrees() and drainVDOPageCache() to be reversed in
initiate drain, the problem with this call in drainZoneTrees() would
become a problem here instead.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>