142637 by sclafani on 2020/02/25 (pair: corwin, raeburn) Gave admin_completion an explicit struct vdo reference.
This started with trying to adopt the kernel's container_of()
macro. In reviewing a larger change that did so, Ken pointed out
there was at least one place where it was unsafe to remove the
offsetof==0 static assertion because we're implicitly depending
elsewhere on the container and field to share addresses.

In trying to find places with those implicit casts, I noticed that
in a bunch of places we're passing admin_completion's "completion"
parent as VDO pointer without having to cast it. The parent field
is a void *, so can be passed to any pointer parameter. There's no
need to use parent for that since everything that depends on
completion->parent being a struct vdo reference already has a
reference to the enclosing admin_completion.

.../adminCompletion.h
.../adminCompletion.c
Added a VDO pointer to struct admin_completion to use instead of
stashing it in the completion's parent field and having to unsafely
(and often implicitly) cast it back.
Respelled a cast to use container_of(), which is both a safer idiom and
necessary since "completion" is no longer the first field.
Made prepare_admin_sub_task_on_thread() pass the address of
"completion" as the parent, not the admin_completion. The user of
that parent expects it to be a vdo_completion. This worked before
because they used to be the same address.
Fixed misindentations in initialize_admin_completion().

.../adminCompletion.c
.../vdoResize.c
.../vdoResizeLogical.c
.../vdoResume.c
.../vdoSuspend.c
Switched to using the new vdo field.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>