142661 by sclafani on 2020/02/26 (pair: raeburn) Respelled cast idioms using the kernel's container_of() macro [VDOUP-6]
.../actionManager.c
.../blockMapRecovery.c
.../flush.c
.../logicalZone.c
.../partitionCopy.c
.../readOnlyNotifier.c
.../readOnlyRebuild.c
.../referenceCountRebuild.c
Respelled direct casts that first asserted offsetof() == 0 with
container_of(), but only in places where NULL is implicitly checked.
They happen to all be completion pointers. In all cases, I manually
tested with a non-zero offset to confirm that there are
no hidden assumptions elsewhere that the offset is zero, and
also attempted to verify by inspection.

.../referenceCountRebuild.c
Fixed two places where the address of the completion field should be
taken, but the address of the containing structure was being used.

.../blockMapRecovery.c
Narrowed the argument type of freeRecoveryCompletion() to the
type of the structure being freed, as is conventional, simplifying
the error paths in the constructor.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>