142665 by jwiele on 2020/02/26 (pair: sclafani) Code cleanup related to UDS change 142624 [UDS-3031].
Change 142624 simplified logger initialization as a followup
to change 140037. openLogger was obsoleted by change
140037. After changes 140037 and 142624, logger initialization
in user space is done by an implicit call to initLogger the
first time a message is logged and logger cleanup is done at
process exit.

.../vdo/kernel/logger.h
Remove the comment about the (lack of) thread safety of
openLogger and closeLogger since they do not exist in kernel
space. Kernel logging is done using printk, which needs no
special initialization or cleanup and takes care of its own
thread safety.

Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>