142670 by sclafani on 2020/02/27 (pair: sweettea) Respelled cast idioms using the kernel's container_of() macro [VDOUP-6]
.../blockMapTree.c
.../forest.c
Respelled direct casts that first asserted offsetof() == 0 with
container_of(), but only in places where NULL is implicitly checked.
They happen to all be waiter pointers. In all cases, I manually
tested with a non-zero offset to confirm that there are
no hidden assumptions elsewhere that the offset is zero, and
also attempted to verify by inspection.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>