142678 by jwiele on 2020/02/27 (pair: sclafani) Merge 142624 from linux-vdo to Krusty. 142624 by jwiele on 2020/02/24 (pair: sweettea) ALB-3031 Simplify user space logger, removing ref count.
The openLogger method contains a reference counter that has
been rendered mostly obsolete by change 140037. Its only
remaining purpose is to ensure that logger initialization is
only performed once. It is only incremented and never
decremented in product code. openLogger can be simplified by
eliminating the reference counter and using a performOnce
method to do the initialization.

Historically, all UDS API functions called udsInitialize to
set up global UDS state for the process before doing any UDS
operations. udsShutdown would tear down the global
state. Typically, udsShutdown would be called by an exit
handler, but it could be called explicitly by an application.

udsInitialize called openLogger to initialize the global
logging state, including opening a log file and/or syslog
socket and setting global logging configuration. As part of
tear-down, udsShutdown called closeLogger to close any open
files or sockets and clean up any data associated with the
logger state.

The logger contained a reference counter so that the
initialization would only be performed the first time
openLogger was called and cleanup would be performed when
closeLogger decremented the reference counter to zero.

As of change 140037, the last bit of global state other than
logging had been eliminated. For that reason, udsInitialize
and udsShutdown were removed. The call to openLogger was moved
to logMessagePack so that the logger would be initialized the
first time an application process logged a message. The call
to closeLogger was removed since there was no place to put it
in the absence of udsShutdown and logger teardown was left to
process exit.

.../logger.h
Eliminate closeLogger since it is only used by test code that
needs to manipulate the logging configuration.
Replace closeLogger with a test-only reinitLogger function for
tests of logging that need to manipulate the logging
configuration.

142680 by sclafani on 2020/02/27 (pair: raeburn)
Respelled a cast using the kernel's container_of() macro [VDOUP-6]

.../extent.h
Respelled a direct cast that first asserted offsetof() == 0 with
container_of(). NULL is implicitly checked. I manually tested
with a non-zero offset to confirm that there are no hidden
assumptions elsewhere that the offset is zero, and also
attempted to verify by inspection.

.../extent.c
Fixed two places where the address of the completion field should be
taken, but the address of the containing structure was being used.

Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>