142690 by sclafani on 2020/02/28 (pair: msakai) Respelled cast idioms using the kernel's container_of() macro [VDOUP-6]
.../vdoRecoveryInternals.h
.../vio.h
Respelled two direct casts that first asserted offsetof() == 0 with
container_of(), but only in places where NULL is implicitly checked.
They happen to be completion pointers. In both cases, I manually
tested with a non-zero offset to confirm that there are
no hidden assumptions elsewhere that the offset is zero, and
also attempted to verify by inspection.

.../vdoRecovery.c
Fixed places where the address of the completion field should be
taken, but the address of the containing structure was being used.
Eliminated an unnecessary local variable.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>