143096 by jshimkus on 2020/03/30 (pair: joe) Rename TraceLocationRecord to TraceLocation. (VDOUP-6)
As a preliminary step in modifying the trace.h types to comport with
kernel standards this changeset deletes the existing TraceLocation type
and renames the TraceLocationRecord type to TraceLocation so that
eventually detypedefing will maintain the semantic concordance between
type and usage.

.../trace.h
Renamed TraceLocationRecord to TraceLocation.

.../allocatingVIO.h
.../dataVIO.h
.../trace.c
.../vio.h
.../dataKVIO.h
.../ioSubmitter.c
.../kvio.h
.../verify.c
Adapt to the above changes.

Co-authored-by: jshimkus <jshimkus@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>