143116 by jshimkus on 2020/03/31 (pair: raeburn, scliff) Detypedef trace.h (VDOUP-6)
In detypdefing TraceLocation we lose the ability to specify the
constant nature of the struct as part of the type.  Consequently, the
code has had to be additionally modified to explicitly include the
const specifier everywhere 'struct trace_location' is used.

.../trace.h
Detypedefed Trace, TraceLocation and TraceRecord.

.../trace.c
.../allocatingVIO.h
.../dataVIO.h
.../vio.h
.../dataKVIO.h
.../ioSubmitter.c
.../ktrace.c
.../ktrace.h
.../kvio.c
.../kvio.h
.../verify.c
Adapt to the above changes.

Co-authored-by: jshimkus <jshimkus@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>