143229 by sweettea on 2020/04/06 (pair: raeburn) Make dedupeIndex use a normal work item.
We have to log on a different thread, but we can bundle
a bunch of reports into a single work queue item, and
then ratelimit the reporting. In this way, we can avoid
using enqueue_work_queue_delayed(), which adds a lot of
complication to workqueue (and whose only user is this).

.../dedupeIndex.c
Changed to report immediately after timing out a batch
of requests.

Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>