143337 by raeburn on 2020/04/13 (pair: sclafani) Merge 143315 to krusty. 143315 by raeburn on 2020/04/10 (pair: sclafani) Restore dormancy/timeout adjustment code path.
Change 141488 updated the wait calls to use interruptible
waits. Unfortunately, the check for interrupts, intended to
bypass the timeout adjustment code path for that case, also
catches the timeout case as well. This results in the worker
thread never adjusting the timeout, and never switching to
"dormant" mode once woken up.

This is visible as the worker threads maintaining a low level
of CPU utilization (e.g., around 1% on one of my test
machines) even when no requests have come in for an extended
period of time. The original intent of the code was that it
would have switched to "dormant" mode, wherein the next
incoming request would've triggered waking the worker thread.

It also appears that kthread starts threads with all signals
ignored, so unless we change that, we won't receive signals
anyway. This makes the interrupt check unnecessary.

.../requestQueueKernel.c
Mostly restore the pre-141488 code path for checking the
results of waiting for a request to come in.

143347 by bjohnston on 2020/04/13 (pair: sweettea)
[VDOUP-6] Decamelcase vdoPageCache atomics.

.../vdoPageCache.h
Decamelcase vdoPageCache atomic stats.

.../blockMap.c
.../vdoPageCache.c
Update code to use new names.

Co-authored-by: raeburn <raeburn@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>