143367 by jshimkus on 2020/04/14 (pair: scliff) Merge 143332 to krusty (UDSUP-1). 143329 by jshimkus on 2020/04/13 (pair: sweettea) Detypedef cacheCounters.[ch] (UDSUP-1).
.../cacheCounters.h
.../cacheCounters.c
Detypedef according to kernel standards.

.../pageCache.c
.../sparseCache.c
Adapt to the above changes.

143332 by jshimkus on 2020/04/13 (pair: sweettea)
Decamelcase and reindent cacheCounters.[ch] (UDSUP-1).

.../cacheCounters.h
.../cacheCounters.c
Decamelcase and reindent according to kernel standards.

.../pageCache.c
.../sparseCache.c
Adapt to the above changes.

143371 by jshimkus on 2020/04/14 (pair: scliff)
Merge 143335 to krusty (UDSUP-1).
143335 by jshimkus on 2020/04/13 (pair: sweettea)
Detypedef cachedChapterIndex.[ch] (UDSUP-1).

.../cachedChapterIndex.h
.../cachedChapterIndex.c
Detypedef according to kernel standards.

.../searchList.h
.../searchList.c
.../sparseCache.c
Adapt to the above changes.

143374 by jshimkus on 2020/04/14 (pair: scliff)
Merge 143344 to krusty (UDSUP-1).
143344 by jshimkus on 2020/04/13 (pair: sweettea)
Decamelcase and reindent cachedChapterIndex.[ch] (UDSUP-1).

.../cachedChapterIndex.h
.../cachedChapterIndex.c
Decamelcase and reindent according to kernel standards.

.../searchList.c
.../sparseCache.c
Adapt to the above changes.

143380 by sclafani on 2020/04/14 (pair: raeburn)
Added a compiler.h include in preparation for __must_check switchover.

Using the kernel __must_check macro instead of
__attribute__((warn_unused_result)) requires having access to it,
and there's a file that will need it that doesn't already
include it directly or indirectly. Having the include already be
prepared makes it easier to iterate/revert with a script that does
the conversion. [VDOUP-7]

.../vdoState.h
Included compiler.h.

Co-authored-by: jshimkus <jshimkus@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>