143394 by corwin on 2020/04/15 (pair: sweettea) Eliminate AsyncEnqueueable [VDOUP-3].
The first real step in eliminating enqueueables is to modify AsyncLayer
to not use them. This change does so by embedding a kvdo_work_item in
completion, and defining a version of that structure for use in user
space.

.../completion.h
Added a definition of struct kvdo_work_item which, for now, simply
contains a single FunnelQueueEntry. This definition applies when
compiled with -DUSER_SPACE. Otherwise, workQueue.h is included to
provide the appropriate definition.
Added a 'work_item' field to completion which is a kvdo_work_item.

.../physicalLayer.h
Modified enqueuer() to take a completion instead of an enqueueable.

.../completion.c
Reflected change to enqueuer(). The assertion used to catch attempts to
enqueue completions without enqueueables has moved into the kernel
layer's implementation of the enqueuer() method.

.../adminCompletion.c
Reflected change to enqueuer().

.../kernelVDO.h
.../kernelVDO.c
Modified kvdo_enqueue() to take a completion instead of an enqueueable
in accordance with the change to enqueuer().

.../kvio.c
Added ENQUEUEABLE_REMOVAL ifdef around the assertion that
metadata_kvios are not larger than 256 bytes. With the addition of
the work_item field to vdo_completion, this is no longer true. This
growth will be reversed in the next change which removes the
enqueueable field from kvio. In the meantime, an extra assertion has
been added in the else arm of the ifdef. This assertion will fire if
the struct enqueueable type goes away, signaling that this ifdef
should be removed.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>