143431 by sclafani on 2020/04/15 (pair: sweettea) Converted vdo/kernel and vdo/user to use __must_check [VDOUP-7]
This is a script-driven substitution of the kernel's __must_check
macro for all uses in vdo/kernel/* and vdo/user/* of
__attribute__((warn_unused_result)). The majority practice in the
kernel is to place it immediately in front of the function name,
instead of first or last as we have always done, so also
re-arranged declarations to conform. See UDS change 143379 for
more details.

.../batchProcessor.h
.../bufferPool.h
.../dataKVIO.c
.../dataKVIO.h
.../dedupeIndex.h
.../deviceConfig.c
.../deviceConfig.h
.../deviceRegistry.c
.../deviceRegistry.h
.../dmvdo.c
.../errors.h
.../kernelLayer.h
.../kvio.c
.../kvio.h
.../memoryUsage.h
.../threads.h
.../vdoStringUtils.h
.../verify.c
.../workQueue.c
.../workQueueStats.h
Replaced __attribute__((warn_unused_result)) with __must_check.
Moved __must_check to where the kernel usually puts it.
Re-filled modified function declaration, stripping alignment
whitespace and re-indenting parameters using tabs.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>