143475 by jshimkus on 2020/04/17 (pair: joe) Merge 143436 to krusty (UDSUP-1). 143436 by jshimkus on 2020/04/16 (pair: sweettea) Detypedef uds.h (UDSUP-1).
Special note must be made of the changes from the definition of
UdsConfiguration (typedef struct udsConfiguration *UdsConfiguration)
used in "const UdsConfiguration" being modified to
"const struct uds_configuration *".

"const UdsConfiguration" declares a *constant pointer* to a struct; it
does not declare a pointer to a *constant struct*.  Analysis of the
usage indicates that a constant pointer had no value in the code
whereas a const struct made sense in how the struct was used as input
only.  As a result the constant pointer of the previous implementtation
was changed to a pointer to a constant struct.

.../uds.h
Detypedefed according to kernel standards.

.../cachedChapterIndex.h
.../cachedChapterIndex.c
.../chapterIndex.h
.../chapterIndex.c
.../common.h
.../config.h
.../config.c
.../hashUtils.h
.../hashUtils.c
.../index.h
.../index.c
.../indexConfig.c
.../indexLayout.h
.../indexLayout.c
.../indexLayoutLinuxKernel.c
.../indexPageMap.h
.../indexPageMap.c
.../indexSession.h
.../indexSession.c
.../masterIndex005.c
.../masterIndex006.c
.../masterIndexOps.h
.../openChapterZone.h
.../openChapterZone.c
.../recordPage.h
.../recordPage.c
.../request.h
.../sparseCache.h
.../sparseCache.c
.../uds-block.h
.../udsMain.c
.../volume.h
.../volume.c
Adapt to the above changes.

143524 by jshimkus on 2020/04/20 (pair: joe)
Adapt linux-vdo to krusty uds.h detypedef change 143475 (UDSUP-1).

.../dataVIO.h
.../hashLockInternals.h
.../hashZone.h
.../hashZone.c
.../vdo.c
.../vdoInternal.h
.../volumeGeometry.h
.../volumeGeometry.c
.../dataKVIO.h
.../dedupeIndex.c
Adapted to krusty uds.h detypedef.

Co-authored-by: jshimkus <jshimkus@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>