143536 by sweettea on 2020/04/20 (pair: sclafani, raeburn) [VDO-4959] Use kernel work item to report dedupe timeouts.
The existing approach is unsafe, as freeing the dedupe index
does not wait for the events to be reported, and may leave
a work item on a queue or attempt to requeue a work item on
a freed queue. This bug was introduced in @143229.

Additionally, the existing approach puts a work item on the
cpu queue that isn't cpu-intensive; we hope one day to switch
the other types of cpu queue work (hashing and compression)
to use a normal kernel work queue, so moving this work item
off the cpu queue is helpful toward that goal.

This change changes to using a kernel work item instead,
which is easy to cancel when we're freeing the dedupe index.

.../dedupeIndex.c
Use a kernel work item instead of ours to report timeouts.

Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>