143539 by corwin on 2020/04/20 (pair: sweettea, sclafani) Removed kvdo_enqueueable from vdo/kernel [VDOUP-3].
This change converts the kernel to use the work item embedded directly
in vdo_completion instead of using enqueueables when enqueuing
completions.

This change will also fix the failures in MemoryMeasurement [VDO-4956]
which were introduced by 143394 which added the work item to
vdo_completion, but hadn't yet removed the enqueueable it is replacing.

.../kernelLayer.h
Removed declaration of kvdo_enqueueable.

.../kernelLayer.c
Made kvdo_create_enqueueable() a no-op.
Made kvdo_destroy_enqueueable() merely assert that the enqueueable to
be destroyed is already NULL.

.../kernelVDO.c
Modified kvdo_enqueue_work() and kvdo_enqueue() to use the work item
that is embedded in the completion rather than an enqueueable.

.../kvio.h
.../kvio.c
Removed the 'enqueueable' member from struct kvio.
Reimplemented functions which interact with work_items via kvios and
vice versa to use the work_item in the kvio's vio's completion.
Removed the setting up of the enqueueable from initialize_kvio().
Re-enabled metadata_kvio size assertion.

.../ioSubmitter.c
Reflected changes to kvio.
Inverted the kvio_merge NULL check in get_mergeable_locked() in order
to reduce a level of identation.

.../dataKVIO.h
.../dataKVIO.c
.../dedupeIndex.c
Reflected changes to kvio.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>